refactor(cli): split x manager into bootstrap/cli package
Move every Click command group out of x.py into per-group register_* modules under bootstrap/cli, backed by a shared runtime module for dry-run state and common helpers. Drop unused module-level helpers. x.py is now a thin entrypoint.
