+++ bootstrap/cli/generate.py
@@ -1,0 +1,265 @@
+from __future__ import annotations
+
+import asyncio
+import shutil
+
+from pathlib import Path
+
+import click
+
+from click_aliases import ClickAliasedGroup
+from colorama import Fore
+from colorama import Style
+from watchfiles import awatch
+
+from bootstrap.cli import runtime
+from bootstrap.color import styled
+from bootstrap.constant import I18N_ROOT
+from bootstrap.constant import PROJECT_ROOT
+from bootstrap.constant import PROTOBUF_DART_OUT_PATH
+from bootstrap.constant import PROTOBUF_PYTHON_OUT_PATH
+from bootstrap.constant import PROTOBUF_SCHEMA_PATH
+from bootstrap.data.codegen import CODEGEN_DART
+from bootstrap.log import info
+from bootstrap.log import warning
+from bootstrap.utils import get_command
+
+
+def _run_protobuf() -> None:
+    protoc = get_command("protoc")
+
+    total = 0
+    failed: set[str] = set()
+
+    if not PROTOBUF_PYTHON_OUT_PATH.exists():
+        warning("Python protobuf output path not found, creating it.")
+        PROTOBUF_PYTHON_OUT_PATH.mkdir(parents=True, exist_ok=True)
+    if not PROTOBUF_DART_OUT_PATH.exists():
+        warning("Dart protobuf output path not found, creating it.")
+        PROTOBUF_DART_OUT_PATH.mkdir(parents=True, exist_ok=True)
+
+    for file in PROTOBUF_SCHEMA_PATH.glob("*.proto"):
+        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Generating protobuf code for: ") + f"{file}")
+        runtime.execute(
+            [
+                protoc,
+                f"--proto_path={PROTOBUF_SCHEMA_PATH}",
+                f"--python_out={PROTOBUF_PYTHON_OUT_PATH}",
+                f"--dart_out={PROTOBUF_DART_OUT_PATH}",
+                file.name,
+            ],
+            "PROTOBUF CODEGEN OUTPUT",
+        )
+        total += 1
+
+    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Protobuf code generation completed."))
+    if len(failed) == 0:
+        click.echo(styled([Style.BRIGHT, Fore.GREEN], "All files generated successfully."))
+    else:
+        click.echo(
+            styled(Fore.GREEN, "Successfully generated: ")
+            + styled([Style.BRIGHT, Fore.GREEN], f"{total - len(failed)}")
+            + Fore.GREEN
+            + f" file{'s' if total - len(failed) > 1 else ''}."
+        )
+        click.echo(
+            styled(Fore.RED, "Failed to generate: ")
+            + styled([Style.BRIGHT, Fore.RED], f"{len(failed)}")
+            + Fore.RED
+            + f" file{'s' if len(failed) > 1 else ''}: "
+            + ", ".join(failed)
+            + "."
+        )
+
+
+def _run_rust() -> None:
+    native_output_dir = PROJECT_ROOT / "lib" / "native"
+    if native_output_dir.exists():
+        info(f"Removing existing native output directory: {native_output_dir}")
+        shutil.rmtree(native_output_dir)
+    flutter_rust_bridge_codegen = get_command("flutter_rust_bridge_codegen")
+    click.echo(
+        styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
+        + "flutter_rust_bridge_codegen generate"
+    )
+    runtime.execute([flutter_rust_bridge_codegen, "generate"], "FRB CODEGEN OUTPUT")
+    click.echo(
+        styled([Style.BRIGHT, Fore.GREEN], "Rust bridge code generation completed successfully.")
+    )
+
+
+def _run_dart(watch: bool) -> None:
+    click.echo(
+        styled([Style.BRIGHT, Fore.GREEN], "Executing codegen: "),
+    )
+    for codegen in CODEGEN_DART:
+        for file in codegen():
+            click.echo(f"  Modified {file}")
+
+    flutter = get_command("flutter")
+    click.echo(
+        styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
+        + f"flutter pub run build_runner {'watch' if watch else 'build'} --delete-conflicting-outputs"
+    )
+    runtime.execute(
+        [
+            flutter,
+            "pub",
+            "run",
+            "build_runner",
+            "watch" if watch else "build",
+            "--delete-conflicting-outputs",
+        ],
+        "DART BUILDRUNNER OUTPUT",
+    )
+    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Dart build runner completed successfully."))
+
+
+def _run_l10n_once() -> None:
+    flutter = get_command("flutter")
+    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter gen-l10n")
+    runtime.execute([flutter, "gen-l10n"], "FLUTTER GEN-L10N OUTPUT")
+
+
+def register_generate_commands(cli_group: click.Group) -> None:
+    @cli_group.group(aliases=["gen"], cls=ClickAliasedGroup)
+    @click.option(
+        "--format",
+        "-f",
+        "format_source",
+        is_flag=True,
+        default=False,
+        help="Run formatter after generation.",
+    )
+    @click.pass_context
+    def generate(ctx: click.Context, format_source: bool):
+        """Code generation related commands."""
+        ctx.ensure_object(dict)
+        ctx.obj["format_source"] = format_source
+
+    @generate.command("all")
+    @click.pass_context
+    def all_cmd(ctx: click.Context):
+        """Generate all code."""
+        _run_protobuf()
+        _run_rust()
+        _run_dart(watch=False)
+        _run_l10n_once()
+
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.command(aliases=["proto", "pb"])
+    @click.pass_context
+    def protobuf(ctx: click.Context):
+        """Generate protobuf code for all supported languages."""
+        _run_protobuf()
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.command("rust", aliases=["rs"])
+    @click.pass_context
+    def rust_cmd(ctx: click.Context):
+        """Generate flutter-rust-bridge glue code."""
+        _run_rust()
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.command("dart")
+    @click.option("--watch", "-w", is_flag=True, default=False, help="Run in watch mode.")
+    @click.pass_context
+    def dart_build_runner(ctx: click.Context, watch: bool):
+        """Run `flutter pub run build_runner build`."""
+        _run_dart(watch=watch)
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.command("l10n")
+    @click.option("--watch", "-w", is_flag=True, default=False, help="Run in watch mode.")
+    @click.pass_context
+    def gen_l10n(ctx: click.Context, watch: bool):
+        """Generate localization files."""
+        if watch:
+
+            async def watch_l10n():
+                _run_l10n_once()
+                async for _ in awatch(str(I18N_ROOT)):
+                    _run_l10n_once()
+
+            try:
+                asyncio.run(watch_l10n())
+            except KeyboardInterrupt:
+                click.echo(styled([Style.BRIGHT, Fore.YELLOW], "\nWatch mode interrupted by user."))
+                return
+
+        _run_l10n_once()
+        click.echo(
+            styled([Style.BRIGHT, Fore.GREEN], "Localization generation completed successfully.")
+        )
+
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.group("values", cls=ClickAliasedGroup)
+    def generate_values():
+        """Generate value-dependent code from the selected workspace."""
+
+    @generate_values.command("dogma-units")
+    @click.pass_context
+    def dogma_units_cmd(ctx: click.Context):
+        """Generate dogma unit ID constants."""
+        from bootstrap.data.codegen.dogma_unit_id import codegen_dart
+
+        files = asyncio.run(codegen_dart(runtime.current_workspace_descriptor()))
+        for file in files:
+            click.echo(f"  Modified {file}")
+
+        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Dogma unit ID generation completed."))
+
+        if ctx.obj.get("format_source", False):
+            runtime.run_format()
+
+    @generate.command("schema")
+    @click.option(
+        "--dir",
+        "build_dir",
+        type=click.Path(path_type=Path),
+        required=True,
+        help="Build directory containing workspace output.",
+    )
+    @click.option(
+        "--server",
+        "server_id",
+        required=True,
+        help="Server ID for the checkout catalog (e.g., 'serenity').",
+    )
+    @click.option(
+        "--schema-root",
+        type=click.Path(path_type=Path),
+        default=None,
+        help="Unified schema root directory (default from dev config).",
+    )
+    @click.option("--author", default=None, help="Author identifier for the snapshot.")
+    @click.option("--description", default=None, help="Description for the snapshot.")
+    def generate_schema_cmd(build_dir, server_id, schema_root, author, description):
+        """Generate a V2 schema checkout from workspace build output."""
+        import bootstrap.config
+
+        from bootstrap.data.workspace.generate.schema import generate_schema_checkout
+
+        if schema_root is None:
+            bootstrap.config.DeveloperConfiguration.ensure_loaded()
+            schema_root = bootstrap.config.DEV_CONFIGURATION.paths.schema_dir
+
+        hash_ = generate_schema_checkout(
+            config=None,
+            build_dir=build_dir,
+            schema_root=schema_root,
+            server_id=server_id,
+            author=author,
+            description=description,
+        )
+        if hash_:
+            click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Checkout hash: {hash_}"))
+        else:
+            click.echo(styled([Style.BRIGHT, Fore.RED], "No files found — checkout not generated."))
