+++ x.py
@@ -13,6 +13,11 @@ To synchronize the python environment, execute `uv sync` from the commandline.
 Some of the commands support environment variables to pass parameters,
 but that's not recommended. And the script itself won't load dotenv files.
 Please use the configuration files to configure the tool, or pass parameters directly.
+
+**Architecture**
+`x.py` is intentionally tiny: it only bootstraps the environment, builds the
+root Click group, and delegates every command group to a `register_*` function
+under `bootstrap.cli`. Add or modify commands there, not here.
 """
 
 # ruff: noqa: E402
@@ -20,37 +25,14 @@ Please use the configuration files to configure the tool, or pass parameters dir
 
 from __future__ import annotations
 
-import asyncio
-import hashlib
-import json
-import os
-import shutil
-import subprocess
 import sys
-import time
-import zipfile
 
 from pathlib import Path
-from typing import TYPE_CHECKING
-from urllib.parse import unquote
-from urllib.request import urlopen
 
-import click
-
-from click_aliases import ClickAliasedGroup
-from colorama import Fore
-from colorama import Style
-from colorama import init
 from dotenv import load_dotenv
-from watchfiles import awatch
 
-from bootstrap.constant import DEV_CONFIG_PATH
-from bootstrap.constant import I18N_ROOT
-from bootstrap.constant import NATIVE_LIB_ROOT
-from bootstrap.constant import PROJECT_ROOT
-from bootstrap.data.codegen import CODEGEN_DART
-from bootstrap.etc.codeart import generate_codeart
-from bootstrap.remote.channel import Channel
+
+PROJECT_ROOT = Path(__file__).parent.resolve()
 
 
 def __fix_env():
@@ -60,36 +42,18 @@ def __fix_env():
 
 __fix_env()
 
-import bootstrap.config
+import click
 
-from bootstrap.ci.commands import register_ci_commands
-from bootstrap.ci.lint import run_lint as _ci_lint
+from click_aliases import ClickAliasedGroup
+from colorama import Fore
+from colorama import Style
+from colorama import init
+
+from bootstrap.cli import register_all_commands
+from bootstrap.cli import runtime
 from bootstrap.color import styled
 from bootstrap.config import ProjectConfiguration
 from bootstrap.config import WorkspaceCache
-from bootstrap.constant import PROTOBUF_DART_OUT_PATH
-from bootstrap.constant import PROTOBUF_PYTHON_OUT_PATH
-from bootstrap.constant import PROTOBUF_SCHEMA_PATH
-from bootstrap.log import info
-from bootstrap.log import warning
-from bootstrap.remote import SessionManager
-from bootstrap.remote import SessionManagerCommittedError
-from bootstrap.remote import SessionManagerInvalidError
-from bootstrap.remote.session_model import Session
-from bootstrap.remote.session_model import SessionExistsError
-from bootstrap.remote.session_model import SessionStore
-from bootstrap.remote.verify import Verifier
-
-
-if TYPE_CHECKING:
-    from bootstrap.remote.sync import SyncResult
-from datetime import UTC
-
-from bootstrap.data.workspace.config import WorkspaceConfig
-from bootstrap.utils import execute_command
-from bootstrap.utils import get_bin_size
-from bootstrap.utils import get_command
-from bootstrap.utils import get_file_sha1
 
 
 init(autoreset=True)
@@ -105,356 +69,6 @@ ProjectConfiguration.load_from_global()
 WorkspaceCache.load_from_global()
 
 
-DRY_RUN = False
-
-
-def __execute_command(
-    cmd: list, title: str, capture_stdout: bool = False, live_stdout: bool = False
-) -> str:
-    global DRY_RUN
-
-    return execute_command(cmd, title, DRY_RUN, capture_stdout, live_stdout)
-
-
-def __resolve_dev_path(path: Path) -> Path:
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    if path.is_absolute():
-        return path
-    return bootstrap.config.DEV_CONFIGURATION.paths.root / path
-
-
-def __wait_for_http(url: str, timeout_seconds: float = 20.0) -> None:
-    deadline = time.monotonic() + timeout_seconds
-    last_error: Exception | None = None
-    while time.monotonic() < deadline:
-        try:
-            with urlopen(url, timeout=1.0):
-                return
-        except Exception as exception:
-            last_error = exception
-            time.sleep(0.25)
-
-    message = f"Timed out waiting for {url}"
-    if last_error is not None:
-        message += f": {last_error}"
-    raise click.ClickException(message)
-
-
-def __run_foreground(process: subprocess.Popen[str], interrupted_message: str) -> None:
-    try:
-        return_code = process.wait()
-    except KeyboardInterrupt:
-        process.terminate()
-        try:
-            process.wait(timeout=10)
-        except subprocess.TimeoutExpired:
-            process.kill()
-            process.wait()
-        click.echo(styled([Style.BRIGHT, Fore.YELLOW], interrupted_message))
-        return
-
-    if return_code != 0:
-        raise click.ClickException(f"Remote mock process exited with status {return_code}.")
-
-
-def __validate_remote_resource_root(resource_root: str) -> str:
-    normalized = resource_root.strip().strip("/")
-    parts = normalized.split("/")
-    if (
-        not normalized
-        or resource_root.strip().startswith("/")
-        or ".." in parts
-        or any(part == "" for part in parts)
-        or "://" in normalized
-    ):
-        raise click.ClickException(f"Invalid remote resource root: {resource_root!r}")
-    return normalized
-
-
-def __validate_remote_channel(channel: str) -> Channel:
-    normalized = channel.strip()
-    if not normalized or "/" in normalized or ".." in normalized or "%2e" in normalized.lower():
-        raise click.ClickException(f"Invalid remote channel: {channel!r}")
-    try:
-        return Channel(normalized)
-    except ValueError:
-        raise click.ClickException(
-            f"Unknown remote channel: {channel!r}. "
-            f"Expected one of: {', '.join(c.value for c in Channel)}"
-        ) from None
-
-
-def __validate_remote_document_id(document_id: str) -> str:
-    normalized = document_id.strip()
-    if not normalized or "/" in normalized or ".." in normalized or ".." in unquote(normalized):
-        raise click.ClickException(f"Invalid remote document id: {document_id!r}")
-    return normalized
-
-
-def __validate_remote_artifact_id(artifact_id: str) -> str:
-    normalized = artifact_id.strip()
-    if (
-        not normalized
-        or normalized.startswith("-")
-        or "/" in normalized
-        or ".." in normalized
-        or ".." in unquote(normalized)
-        or any(character.isspace() for character in normalized)
-    ):
-        raise click.ClickException(f"Invalid remote bundle artifact id: {artifact_id!r}")
-    return normalized
-
-
-def __read_current_app_version() -> str:
-    pubspec_path = PROJECT_ROOT / "pubspec.yaml"
-    for line in pubspec_path.read_text(encoding="utf-8").splitlines():
-        stripped = line.strip()
-        if stripped.startswith("version:"):
-            return stripped.removeprefix("version:").strip()
-    raise click.ClickException(f"Unable to read app version from {pubspec_path}")
-
-
-def __read_json_object(path: Path, default: dict[str, object]) -> dict[str, object]:
-    if not path.exists():
-        return dict(default)
-    if not path.is_file():
-        raise click.ClickException(f"JSON path is not a file: {path}")
-    try:
-        payload = json.loads(path.read_text(encoding="utf-8"))
-    except json.JSONDecodeError as exception:
-        raise click.ClickException(f"Invalid JSON in {path}: {exception.msg}") from exception
-    if not isinstance(payload, dict):
-        raise click.ClickException(f"JSON payload must be an object: {path}")
-    return payload
-
-
-def __write_json_object(path: Path, payload: dict[str, object]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(payload, indent=4, ensure_ascii=False) + "\n", encoding="utf-8")
-
-
-def __file_sha256(path: Path) -> str:
-    hasher = hashlib.sha256()
-    with path.open("rb") as f:
-        for chunk in iter(lambda: f.read(1024 * 1024), b""):
-            hasher.update(chunk)
-    return hasher.hexdigest()
-
-
-def __read_zip_json(zip_path: Path, member_name: str) -> dict[str, object]:
-    if not zip_path.is_file():
-        raise click.ClickException(f"Bundle archive does not exist: {zip_path}")
-    try:
-        with zipfile.ZipFile(zip_path) as archive, archive.open(member_name) as f:
-            payload = json.loads(f.read().decode("utf-8"))
-    except KeyError as exception:
-        raise click.ClickException(
-            f"Bundle archive is missing {member_name}: {zip_path}"
-        ) from exception
-    except UnicodeDecodeError as exception:
-        raise click.ClickException(
-            f"Bundle archive {member_name} is not valid UTF-8: {zip_path}"
-        ) from exception
-    except json.JSONDecodeError as exception:
-        raise click.ClickException(
-            f"Invalid JSON in bundle archive {member_name}: {zip_path}: {exception.msg}"
-        ) from exception
-    except zipfile.BadZipFile as exception:
-        raise click.ClickException(f"Invalid bundle archive: {zip_path}") from exception
-    if not isinstance(payload, dict):
-        raise click.ClickException(
-            f"Bundle archive {member_name} must be a JSON object: {zip_path}"
-        )
-    return payload
-
-
-def __validate_mc_target_segment(value: str, label: str) -> str:
-    normalized = value.strip()
-    if (
-        not normalized
-        or normalized.startswith("-")
-        or "/" in normalized
-        or ".." in normalized
-        or any(character.isspace() for character in normalized)
-    ):
-        raise click.ClickException(f"Invalid remote publish {label}: {value!r}")
-    return normalized
-
-
-def __execute_command_redacted(cmd: list[str], redacted_cmd: list[str], title: str) -> None:
-    if DRY_RUN:
-        info(f"[Dry-Run] {title}: " + " ".join(redacted_cmd))
-        return
-
-    out = subprocess.run(cmd, capture_output=True, text=True, encoding="utf-8", errors="replace")
-    if out.returncode != 0:
-        message = f"Failed to execute command [{out.returncode}]: " + " ".join(redacted_cmd)
-        stderr = (out.stderr or "").strip()
-        if stderr:
-            message += f"\n{stderr}"
-        raise click.ClickException(message)
-
-
-def __publish_optional_tree(
-    mc: str, source: Path, target: str, *, attrs: dict[str, str] | None = None
-) -> None:
-    if not source.exists():
-        warning(f"Remote publish source tree does not exist, skipping: {source}")
-        return
-    if not source.is_dir():
-        raise click.ClickException(f"Remote publish source tree is not a directory: {source}")
-    target_base = target.rstrip("/")
-    for f in sorted(source.rglob("*")):
-        if not f.is_file():
-            continue
-        rel = f.relative_to(source)
-        remote = f"{target_base}/{rel}"
-        __publish_optional_file(mc, f, remote, attrs=attrs)
-
-
-def __publish_optional_file(
-    mc: str, source: Path, target: str, *, attrs: dict[str, str] | None = None
-) -> None:
-    if not source.exists():
-        warning(f"Remote publish source file does not exist, skipping: {source}")
-        return
-    if not source.is_file():
-        raise click.ClickException(f"Remote publish source path is not a file: {source}")
-    cmd = [mc, "cp"]
-    if attrs:
-        for k, v in attrs.items():
-            cmd.extend(["--attr", f"{k}={v}"])
-    cmd.extend([str(source), target])
-    __execute_command(cmd, "REMOTE PUBLISH")
-
-
-def __publish_remote_origin_to_s3(
-    *,
-    source_dir: Path,
-    endpoint: str,
-    bucket: str,
-    access_key: str,
-    secret_key: str,
-    alias_name: str,
-    resource_root: str,
-    channel: Channel,
-    generation: str,
-    public_download: bool,
-    target: str = "minio",
-) -> None:
-    if not source_dir.exists():
-        raise click.ClickException(f"Remote publish source directory does not exist: {source_dir}")
-    if not source_dir.is_dir():
-        raise click.ClickException(f"Remote publish source path is not a directory: {source_dir}")
-    if not endpoint.strip():
-        raise click.ClickException("Remote publish endpoint must not be empty.")
-    if not access_key:
-        raise click.ClickException("Remote publish access key must not be empty.")
-    if not secret_key:
-        raise click.ClickException("Remote publish secret key must not be empty.")
-    if not generation or not generation.strip():
-        raise click.ClickException("Remote publish generation must not be empty.")
-
-    resolved_bucket = __validate_mc_target_segment(bucket, "bucket")
-    resolved_alias = __validate_mc_target_segment(alias_name, "alias")
-    resolved_resource_root = __validate_remote_resource_root(resource_root)
-    root_dir = source_dir / resolved_resource_root
-    channel_dir = root_dir / "channels" / channel.value
-    gen_dir = channel_dir / ".generations" / generation
-    index_path = gen_dir / "index.json"
-
-    if not index_path.exists() or not index_path.is_file():
-        if not channel_dir.is_dir():
-            raise click.ClickException(
-                f"Remote publish channel directory does not exist: {channel_dir}"
-            )
-        gen_dir.mkdir(parents=True, exist_ok=True)
-        for item in channel_dir.iterdir():
-            if item.name == ".generations":
-                continue
-            dst = gen_dir / item.name
-            if item.is_dir():
-                if not dst.exists():
-                    shutil.copytree(item, dst)
-            elif not dst.exists():
-                shutil.copy2(item, dst)
-
-    mc = get_command("mc")
-    bucket_target = f"{resolved_alias}/{resolved_bucket}"
-    redacted = "<redacted>"
-    __execute_command_redacted(
-        [mc, "alias", "set", resolved_alias, endpoint, access_key, secret_key, "--api", "s3v4"],
-        [mc, "alias", "set", resolved_alias, endpoint, redacted, redacted, "--api", "s3v4"],
-        "REMOTE PUBLISH ALIAS",
-    )
-    if target == "minio":
-        __execute_command([mc, "mb", "--ignore-existing", bucket_target], "REMOTE PUBLISH")
-    if target == "minio":
-        if public_download:
-            __execute_command([mc, "anonymous", "set", "download", bucket_target], "REMOTE PUBLISH")
-        else:
-            __execute_command([mc, "anonymous", "set", "none", bucket_target], "REMOTE PUBLISH")
-
-    target_root = f"{bucket_target}/{resolved_resource_root}"
-    # Step 1: mirror shared content (idempotent, safe to interrupt)
-    __publish_optional_tree(
-        mc,
-        root_dir / "documents" / "body",
-        f"{target_root}/documents/body",
-        attrs={
-            "Cache-Control": "immutable, max-age=31536000",
-            "Content-Type": "text/markdown; charset=utf-8",
-        },
-    )
-    __publish_optional_tree(
-        mc,
-        root_dir / "bundles",
-        f"{target_root}/bundles",
-        attrs={"Cache-Control": "immutable, max-age=31536000"},
-    )
-
-    # Step 2: mirror generation catalog tree (tiny JSONs, NEW paths)
-    channel_attrs = {"Cache-Control": "max-age=300", "Content-Type": "application/json"}
-    __publish_optional_tree(
-        mc,
-        gen_dir / "documents",
-        f"{target_root}/channels/{channel}/.generations/{generation}/documents",
-        attrs=channel_attrs,
-    )
-    __publish_optional_tree(
-        mc,
-        gen_dir / "bundles",
-        f"{target_root}/channels/{channel}/.generations/{generation}/bundles",
-        attrs=channel_attrs,
-    )
-    __publish_optional_file(
-        mc,
-        gen_dir / "app" / "releases.json",
-        f"{target_root}/channels/{channel}/.generations/{generation}/app/releases.json",
-        attrs=channel_attrs,
-    )
-
-    # Step 3: atomic commit — copy generation's index.json to live path
-    __publish_optional_file(
-        mc,
-        index_path,
-        f"{target_root}/channels/{channel}/index.json",
-        attrs={"Cache-Control": "no-cache", "Content-Type": "application/json"},
-    )
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Uploaded remote origin: ") + str(source_dir))
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Target bucket: ") + bucket_target)
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Generation: ") + generation)
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Remote index URL: ")
-        + (
-            f"{endpoint.rstrip('/')}/{resolved_bucket}"
-            f"/{resolved_resource_root.strip('/')}"
-            f"/channels/{channel.value}/index.json"
-        )
-    )
-
-
 @click.group(
     context_settings={
         "help_option_names": ["-h", "--help"],
@@ -465,4418 +79,13 @@ def __publish_remote_origin_to_s3(
 @click.option("--workspace", "--ws", "ws_name", default=None, help="Set current workspace.")
 def cli(dry_run, ws_name):
     """EFA Workspace Manager."""
-    global DRY_RUN
-    DRY_RUN = dry_run
+    runtime.set_dry_run(dry_run)
 
     if ws_name:
         WorkspaceCache.select_workspace(ws_name)
 
 
-@cli.command()
-@click.option("--no-check", "no_check", is_flag=True, default=False, help="Skip linting step.")
-@click.option(
-    "--check",
-    "check_only",
-    is_flag=True,
-    default=False,
-    help="Check-only mode: verify without modifying files.",
-)
-@click.option(
-    "--lang",
-    type=click.Choice(["all", "python", "dart", "rust", "site"]),
-    default="all",
-    help="Limit linting to a specific language (default: all).",
-)
-def lint(no_check: bool, check_only: bool, lang: str):
-    """Lint, fix and format code"""
-    _ci_lint(lang, no_check=no_check, check_only=check_only, dry_run=DRY_RUN)
-
-
-@cli.command("format", aliases=["fmt"])
-@click.pass_context
-def format_cmd(ctx: click.Context):
-    """Format the code. This is equivalent to `x lint --no-check`."""
-    ctx.invoke(lint, no_check=True)
-
-
-@cli.group(aliases=["ws"], cls=ClickAliasedGroup)
-def workspace():
-    """Workspace related commands."""
-
-
-@workspace.command("list", aliases=["ls"])
-def list_cmd():
-    """List configured workspaces."""
-    workspaces = bootstrap.config.CONFIGURATION.resources
-    if len(workspaces) == 0:
-        click.echo(
-            styled([Style.BRIGHT + Fore.RED], "Error: ")
-            + styled(Fore.RED, "No workspace configured.")
-        )
-        exit(1)
-
-    click.echo(
-        styled(Fore.GREEN, "Found ")
-        + styled([Style.BRIGHT, Fore.GREEN], f"{len(workspaces)}")
-        + styled(Fore.GREEN, " workspace configurations.")
-    )
-    has_not_found = set()
-    has_warning = set()
-    has_error = set()
-    for ws_key, ws_def in workspaces.items():
-        if ws_def.descriptor.exists():
-            click.echo(
-                styled([Style.BRIGHT, Fore.GREEN], "- [√] ") + f"{ws_key}: {ws_def.descriptor}"
-            )
-            descriptor = WorkspaceConfig.load_from_descriptor(ws_def.descriptor, no_check=True)
-
-            if descriptor.ignore:
-                click.echo(
-                    styled([Style.BRIGHT, Fore.YELLOW], "  [!] Warning: ")
-                    + "workspace is marked as ignored.",
-                )
-                has_warning.add(ws_key)
-                continue
-
-            if not descriptor.resources.fsd.exists() or not descriptor.resources.fsd.is_dir():
-                click.echo(
-                    styled([Style.BRIGHT, Fore.RED], "  [!] Error: ")
-                    + f"FSD path '{descriptor.resources.fsd}' does not exist or is not a directory.",
-                )
-                has_error.add(ws_key)
-
-            if (
-                not descriptor.resources.resource_index.exists()
-                or not descriptor.resources.resource_index.is_file()
-            ):
-                click.echo(
-                    styled([Style.BRIGHT, Fore.RED], "  [!] Error: ")
-                    + f"Resource index '{descriptor.resources.resource_index}' does not exist or is not a file.",
-                )
-                has_error.add(ws_key)
-
-            if (
-                not descriptor.resources.application_index.exists()
-                or not descriptor.resources.application_index.is_file()
-            ):
-                click.echo(
-                    styled([Style.BRIGHT, Fore.RED], "  [!] Error: ")
-                    + f"Application index '{descriptor.resources.application_index}' does not exist or is not a file.",
-                )
-                has_error.add(ws_key)
-
-            if (
-                not descriptor.metadata.start_cfg.exists()
-                or not descriptor.metadata.start_cfg.is_file()
-            ):
-                click.echo(
-                    styled([Style.BRIGHT, Fore.RED], "  [!] Error: ")
-                    + f"Start configuration '{descriptor.metadata.start_cfg}' does not exist or is not a file.",
-                )
-                has_error.add(ws_key)
-        else:
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], "- [!] ")
-                + f"{ws_key}: "
-                + styled([Style.BRIGHT, Fore.RED], "Descriptor not found: ")
-                + f"{ws_def.descriptor}"
-            )
-            has_not_found.add(ws_key)
-
-    if len(has_not_found) > 0:
-        click.echo(
-            styled(Fore.RED, "Missing ")
-            + styled([Style.BRIGHT, Fore.RED], f"{len(has_not_found)}")
-            + styled(
-                Fore.RED,
-                " descriptor" + ("s" if len(has_not_found) > 1 else "") + ": ",
-            )
-            + ", ".join(has_not_found)
-        )
-    if len(has_warning) > 0:
-        click.echo(
-            styled(Fore.YELLOW, "Warning in ")
-            + styled([Style.BRIGHT, Fore.YELLOW], f"{len(has_warning)}")
-            + styled(
-                Fore.YELLOW,
-                " workspace" + ("s" if len(has_warning) > 1 else "") + ": ",
-            )
-            + ", ".join(has_warning)
-        )
-    if len(has_error) > 0:
-        click.echo(
-            styled(Fore.RED, "Error in ")
-            + styled([Style.BRIGHT, Fore.RED], f"{len(has_error)}")
-            + styled(
-                Fore.RED,
-                " workspace" + ("s" if len(has_error) > 1 else "") + ": ",
-            )
-            + ", ".join(has_error)
-        )
-
-    if has_error or has_not_found:
-        exit(1)
-
-
-def __get_workspace(name) -> Path:
-    if not isinstance(name, str):
-        click.echo(styled([Style.BRIGHT, Fore.RED], "Invalid name: ") + f"{name!r}")
-    name = name.strip()
-
-    if len(name) == 0:
-        click.echo(styled([Style.BRIGHT, Fore.RED], "Invalid name: ") + "empty")
-        exit(1)
-
-    workspaces = bootstrap.config.CONFIGURATION.resources
-    ws = workspaces.get(name)
-
-    if ws is None:
-        click.echo(styled([Style.BRIGHT, Fore.RED], "Unknown workspace identifier: ") + name)
-        click.echo("Please check if the workspace is registered in the configuration.")
-        exit(1)
-
-    if not ws.descriptor.exists():
-        click.echo(
-            styled([Style.BRIGHT, Fore.YELLOW], "Warning: ") + f"Descriptor for {name} not found."
-        )
-
-    return ws.descriptor
-
-
-def __get_current_workspace_descriptor() -> WorkspaceConfig:
-    name = bootstrap.config.WORKSPACE_CACHE.current_workspace
-    if not name:
-        click.echo(styled([Style.BRIGHT, Fore.RED], "No workspace selected."))
-        click.echo("Please select a workspace using `x workspace list` and `x workspace default`.")
-        exit(1)
-
-    ws = __get_workspace(name)
-    info(f"Resolving workspace: {name} ({ws})")
-    return WorkspaceConfig.load_from_descriptor(ws)
-
-
-@workspace.command()
-@click.argument("name")
-def default(name: str):
-    """Set default build target resource."""
-    _ = __get_workspace(name)  # check
-
-    ws_cache = bootstrap.config.WORKSPACE_CACHE
-    if ws_cache.default_workspace is not None:
-        click.echo(f"Switch default workspace from {ws_cache.default_workspace} to {name}.")
-    else:
-        click.echo(f"Set default workspace to {name}.")
-    ws_cache.default_workspace = name
-    ws_cache.synchronize()
-
-
-@workspace.command()
-@click.option("--pretty", is_flag=True, default=False, help="Pretty print the JSON output.")
-def inspect_json(pretty: bool):
-    """Resolve the workspace configurations and print in JSON format."""
-    descriptor = __get_current_workspace_descriptor()
-    click.echo(descriptor.model_dump_json(indent=4 if pretty else None))
-
-
-@workspace.command()
-@click.option("--pretty", is_flag=True, default=False, help="Pretty print the JSON output.")
-def cache(pretty: bool):
-    """Print current workspace cache in JSON format."""
-    click.echo(bootstrap.config.WORKSPACE_CACHE.model_dump_json(indent=4 if pretty else None))
-
-
-@cli.group()
-def config():
-    """Configuration related commands."""
-
-
-@config.command()
-def display():
-    """Print loaded configuration in JSON format."""
-    click.echo(bootstrap.config.CONFIGURATION.model_dump_json(indent=4))
-
-
-@cli.group(aliases=["gen"], cls=ClickAliasedGroup)
-@click.option(
-    "--format",
-    "-f",
-    "format_source",
-    is_flag=True,
-    default=False,
-    help="Run formatter after generation.",
-)
-@click.pass_context
-def generate(ctx: click.Context, format_source: bool):
-    """Code generation related commands."""
-    ctx.ensure_object(dict)
-    ctx.obj["format_source"] = format_source
-
-
-@generate.command("all")
-@click.pass_context
-def all_cmd(ctx: click.Context):
-    """Generate all code."""
-    ctx.obj["_in_generate_all"] = True
-    ctx.invoke(protobuf)
-    ctx.invoke(rust_cmd)
-    ctx.invoke(dart_build_runner)
-    ctx.invoke(gen_l10n)
-
-    if ctx.obj.get("format_source", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.command(aliases=["proto", "pb"])
-@click.pass_context
-def protobuf(ctx: click.Context):
-    """Generate protobuf code for all supported languages."""
-    protoc = get_command("protoc")
-
-    total = 0
-    failed = set()
-
-    if not PROTOBUF_PYTHON_OUT_PATH.exists():
-        warning("Python protobuf output path not found, creating it.")
-        PROTOBUF_PYTHON_OUT_PATH.mkdir(parents=True, exist_ok=True)
-    if not PROTOBUF_DART_OUT_PATH.exists():
-        warning("Dart protobuf output path not found, creating it.")
-        PROTOBUF_DART_OUT_PATH.mkdir(parents=True, exist_ok=True)
-
-    for file in PROTOBUF_SCHEMA_PATH.glob("*.proto"):
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Generating protobuf code for: ") + f"{file}")
-        __execute_command(
-            [
-                protoc,
-                f"--proto_path={PROTOBUF_SCHEMA_PATH}",
-                f"--python_out={PROTOBUF_PYTHON_OUT_PATH}",
-                f"--dart_out={PROTOBUF_DART_OUT_PATH}",
-                file.name,
-            ],
-            "PROTOBUF CODEGEN OUTPUT",
-        )
-        total += 1
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Protobuf code generation completed."))
-    if len(failed) == 0:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "All files generated successfully."))
-
-    else:
-        click.echo(
-            styled(Fore.GREEN, "Successfully generated: ")
-            + styled([Style.BRIGHT, Fore.GREEN], f"{total - len(failed)}")
-            + Fore.GREEN
-            + f" file{'s' if total - len(failed) > 1 else ''}."
-        )
-        click.echo(
-            styled(Fore.RED, "Failed to generate: ")
-            + styled([Style.BRIGHT, Fore.RED], f"{len(failed)}")
-            + Fore.RED
-            + f" file{'s' if len(failed) > 1 else ''}: "
-            + ", ".join(failed)
-            + "."
-        )
-
-    if ctx.obj.get("format_source", False) and not ctx.obj.get("_in_generate_all", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.command("rust", aliases=["rs"])
-@click.pass_context
-def rust_cmd(ctx: click.Context):
-    """Generate flutter-rust-bridge glue code."""
-    native_output_dir = PROJECT_ROOT / "lib" / "native"
-    if native_output_dir.exists():
-        info(f"Removing existing native output directory: {native_output_dir}")
-        shutil.rmtree(native_output_dir)
-    flutter_rust_bridge_codegen = get_command("flutter_rust_bridge_codegen")
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
-        + "flutter_rust_bridge_codegen generate"
-    )
-    __execute_command([flutter_rust_bridge_codegen, "generate"], "FRB CODEGEN OUTPUT")
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Rust bridge code generation completed successfully.")
-    )
-
-    if ctx.obj.get("format_source", False) and not ctx.obj.get("_in_generate_all", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.command("dart")
-@click.option("--watch", "-w", is_flag=True, default=False, help="Run in watch mode.")
-@click.pass_context
-def dart_build_runner(ctx: click.Context, watch: bool):
-    """Run `flutter pub run build_runner build`."""
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Executing codegen: "),
-    )
-    for codegen in CODEGEN_DART:
-        for file in codegen():
-            click.echo(f"  Modified {file}")
-
-    flutter = get_command("flutter")
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
-        + f"flutter pub run build_runner {'watch' if watch else 'build'} --delete-conflicting-outputs"
-    )
-    __execute_command(
-        [
-            flutter,
-            "pub",
-            "run",
-            "build_runner",
-            "watch" if watch else "build",
-            "--delete-conflicting-outputs",
-        ],
-        "DART BUILDRUNNER OUTPUT",
-    )
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Dart build runner completed successfully."))
-
-    if ctx.obj.get("format_source", False) and not ctx.obj.get("_in_generate_all", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.command("l10n")
-@click.option("--watch", "-w", is_flag=True, default=False, help="Run in watch mode.")
-@click.pass_context
-def gen_l10n(ctx: click.Context, watch: bool):
-    """Generate localization files."""
-    if watch:
-
-        async def watch_l10n():
-            flutter = get_command("flutter")
-
-            click.echo(
-                styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter gen-l10n"
-            )
-            __execute_command(
-                [flutter, "gen-l10n"],
-                "FLUTTER GEN-L10N OUTPUT",
-            )
-            async for _ in awatch(str(I18N_ROOT)):
-                click.echo(
-                    styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter gen-l10n"
-                )
-                __execute_command(
-                    [flutter, "gen-l10n"],
-                    "FLUTTER GEN-L10N OUTPUT",
-                )
-
-        try:
-            asyncio.run(watch_l10n())
-        except KeyboardInterrupt:
-            click.echo(styled([Style.BRIGHT, Fore.YELLOW], "\nWatch mode interrupted by user."))
-            return
-
-    flutter = get_command("flutter")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter gen-l10n")
-    __execute_command(
-        [flutter, "gen-l10n"],
-        "FLUTTER GEN-L10N OUTPUT",
-    )
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Localization generation completed successfully.")
-    )
-
-    if ctx.obj.get("format_source", False) and not ctx.obj.get("_in_generate_all", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.group("values", cls=ClickAliasedGroup)
-def generate_values():
-    """Generate value-dependent code from the selected workspace."""
-
-
-@generate_values.command("dogma-units")
-@click.pass_context
-def dogma_units_cmd(ctx: click.Context):
-    """Generate dogma unit ID constants."""
-    from bootstrap.data.codegen.dogma_unit_id import codegen_dart
-
-    files = asyncio.run(codegen_dart(__get_current_workspace_descriptor()))
-    for file in files:
-        click.echo(f"  Modified {file}")
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Dogma unit ID generation completed."))
-
-    if ctx.obj.get("format_source", False):
-        ctx.invoke(format_cmd)
-
-
-@generate.command("schema")
-@click.option(
-    "--dir",
-    "build_dir",
-    type=click.Path(path_type=Path),
-    required=True,
-    help="Build directory containing workspace output.",
-)
-@click.option(
-    "--server",
-    "server_id",
-    required=True,
-    help="Server ID for the checkout catalog (e.g., 'serenity').",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Unified schema root directory (default from dev config).",
-)
-@click.option("--author", default=None, help="Author identifier for the snapshot.")
-@click.option("--description", default=None, help="Description for the snapshot.")
-def generate_schema_cmd(
-    build_dir: Path,
-    server_id: str,
-    schema_root: Path | None,
-    author: str | None,
-    description: str | None,
-):
-    """Generate a V2 schema checkout from workspace build output."""
-    from bootstrap.data.workspace.generate.schema import generate_schema_checkout
-
-    if schema_root is None:
-        bootstrap.config.DeveloperConfiguration.ensure_loaded()
-        schema_root = bootstrap.config.DEV_CONFIGURATION.paths.schema_dir
-
-    hash_ = generate_schema_checkout(
-        config=None,
-        build_dir=build_dir,
-        schema_root=schema_root,
-        server_id=server_id,
-        author=author,
-        description=description,
-    )
-    if hash_:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Checkout hash: {hash_}"))
-    else:
-        click.echo(styled([Style.BRIGHT, Fore.RED], "No files found — checkout not generated."))
-
-
-def __env_install():
-    protoc_gen_dart = get_command("protoc-gen-dart")
-    if protoc_gen_dart is None:
-        click.echo(
-            styled([Style.BRIGHT, Fore.RED], "Warning: ") + "protoc-gen-dart not found, installing"
-        )
-        dart = get_command("dart")
-        __execute_command(
-            [dart, "pub", "global", "activate", "protoc_plugin"], "DART ACTIVATE OUTPUT"
-        )
-
-    uv = get_command("uv")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "uv sync")
-    __execute_command([uv, "sync"], "UV SYNC OUTPUT")
-
-    flutter = get_command("flutter")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter pub get")
-    __execute_command([flutter, "pub", "get"], "FLUTTER PUB GET OUTPUT")
-
-    if Path("package.json").exists():
-        pnpm = get_command("pnpm")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "pnpm install")
-        __execute_command([pnpm, "install"], "PNPM INSTALL OUTPUT")
-
-
-def __env_upgrade():
-    uv = get_command("uv")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "uv sync --upgrade")
-    __execute_command([uv, "sync", "--upgrade"], "UV UPGRADE OUTPUT")
-
-    flutter = get_command("flutter")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter pub upgrade")
-    __execute_command([flutter, "pub", "upgrade"], "FLUTTER PUB UPGRADE OUTPUT")
-
-    cargo = get_command("cargo")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "cargo update")
-    __execute_command([cargo, "update"], "CARGO UPDATE OUTPUT")
-
-
-@cli.group(cls=ClickAliasedGroup)
-def dev():
-    """Developer environment commands."""
-
-
-@dev.command("init-cfg")
-def dev_init_cfg():
-    """Create efa.dev.toml from the example template."""
-    if DEV_CONFIG_PATH.exists():
-        click.echo(
-            styled([Style.BRIGHT, Fore.YELLOW], "Warning: ") + "efa.dev.toml already exists."
-        )
-        return
-
-    template = PROJECT_ROOT / "efa.dev.example.toml"
-    shutil.copyfile(template, DEV_CONFIG_PATH)
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Created developer config: ") + str(DEV_CONFIG_PATH)
-    )
-
-
-@dev.group(cls=ClickAliasedGroup)
-def env():
-    """Developer environment setup commands."""
-
-
-@env.command("install")
-def dev_env_install():
-    """Install all tools in the current environment."""
-    __env_install()
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Environment setup completed successfully."))
-
-
-@env.command("upgrade", aliases=["update"])
-def dev_env_upgrade():
-    """Upgrade all tools in the current environment."""
-    __env_upgrade()
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Environment upgrade completed successfully."))
-
-
-@env.command("write-backend")
-def dev_env_write_backend():
-    """Write rust/lib/eve-fit-os/.env from efa.dev.toml."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    native = bootstrap.config.DEV_CONFIGURATION.native
-
-    values = {
-        "FSD_FORMAT": native.fsd_format,
-        "FSD_BINARY_DIR": native.fsd_binary_dir,
-        "FSD_LOC_EN_DIR": native.fsd_loc_en_dir,
-        "OUTPUT_DIR": native.output_dir,
-    }
-    missing = [key for key, value in values.items() if value is None]
-    if len(missing) > 0:
-        raise click.ClickException(
-            "Missing native developer config value(s): " + ", ".join(sorted(missing))
-        )
-
-    env_path = NATIVE_LIB_ROOT / ".env"
-    lines = [f"{key}={value}" for key, value in values.items()]
-    env_path.write_text("\n".join(lines) + "\n", encoding="utf-8")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Wrote backend env: ") + str(env_path))
-
-
-@cli.group(cls=ClickAliasedGroup)
-def remote():
-    """Remote content management — prepare, publish, validate, fetch, mock."""
-
-
-@remote.group("config", cls=ClickAliasedGroup)
-def remote_config():
-    """Remote mock configuration commands."""
-
-
-def __redact_remote_config(config: dict[str, object]) -> dict[str, object]:
-    redacted = dict(config)
-    for sub in ("minio", "s3"):
-        if sub in redacted and isinstance(redacted[sub], dict):
-            sub_dict = dict(redacted[sub])  # type: ignore[arg-type]
-            for key in ("access_key", "secret_key"):
-                if key in sub_dict:
-                    sub_dict[key] = "<redacted>"
-            redacted[sub] = sub_dict
-    return redacted
-
-
-@remote_config.command("display")
-@click.option("--pretty", is_flag=True, default=False, help="Pretty print the JSON output.")
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    default=False,
-    help="Output as machine-readable JSON (no session info).",
-)
-def remote_config_display(pretty: bool, as_json: bool):
-    """Print effective remote developer configuration and current session status."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    bootstrap.config.ProjectConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-    paths = bootstrap.config.DEV_CONFIGURATION.paths
-    origin_path = __resolve_dev_path(remote_cfg.mock_origin_dir)
-
-    resolved: dict[str, str] = {
-        "developerRoot": str(paths.root),
-        "mockOriginPath": str(origin_path),
-    }
-
-    if remote_cfg.minio is not None:
-        minio_data_path = __resolve_dev_path(remote_cfg.minio.data_dir)
-        minio_origin_url = (
-            f"http://{remote_cfg.host}:{remote_cfg.minio.port}/{remote_cfg.minio.bucket}"
-        )
-        resolved["minioDataPath"] = str(minio_data_path)
-        resolved["minioIndexUrl"] = (
-            f"{minio_origin_url.rstrip('/')}"
-            f"/{bootstrap.config.CONFIGURATION.data_schema.resource_root.strip('/')}"
-            f"/channels/{remote_cfg.channel.value}/index.json"
-        )
-
-    payload: dict[str, object] = {
-        "remote": __redact_remote_config(remote_cfg.model_dump(mode="json")),
-        "resolved": resolved,
-    }
-    click.echo(json.dumps(payload, indent=4 if pretty else None))
-
-
-@remote.group("session", cls=ClickAliasedGroup)
-def remote_session():
-    """Staged generation assembly — build, stage, review, commit."""
-
-
-@remote.group("announce", cls=ClickAliasedGroup)
-def remote_announce():
-    """Remote announcement management — sync, author, publish."""
-
-
-def __get_announce_workspace() -> Path:
-    """Get the announcement workspace root path."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    root = bootstrap.config.DEV_CONFIGURATION.paths.root / "announce"
-    return __resolve_dev_path(root)
-
-
-def __resolve_announce_remote_target(
-    target: str | None,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias: str | None,
-) -> tuple[str, str, str, str, str]:
-    """Resolve remote target credentials from CLI args or config."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-
-    if target is None:
-        target = "minio"
-
-    if target == "minio":
-        minio_cfg = remote_cfg.minio
-        if minio_cfg is None:
-            raise click.ClickException(
-                "[remote.minio] is not configured in efa.dev.toml. "
-                "See efa.dev.example.toml for the expected format."
-            )
-        return (
-            endpoint or f"http://{remote_cfg.host}:{minio_cfg.port}",
-            bucket or minio_cfg.bucket,
-            access_key or minio_cfg.access_key,
-            secret_key or minio_cfg.secret_key,
-            alias or minio_cfg.alias,
-        )
-    elif target == "s3":
-        s3_cfg = remote_cfg.s3
-        if s3_cfg is None:
-            raise click.ClickException(
-                "[remote.s3] is not configured in efa.dev.toml. "
-                "See efa.dev.example.toml for the expected format."
-            )
-        return (
-            endpoint or s3_cfg.endpoint,
-            bucket or s3_cfg.bucket,
-            access_key or s3_cfg.access_key,
-            secret_key or s3_cfg.secret_key,
-            alias or s3_cfg.alias,
-        )
-    else:
-        raise click.ClickException(f"Invalid target: {target!r} (expected minio or s3)")
-
-
-@remote_announce.command("sync")
-@click.option(
-    "--target",
-    type=click.Choice(["minio", "s3"]),
-    default=None,
-    help="Remote target (default: minio).",
-)
-@click.option("--endpoint", default=None, help="Override endpoint URL.")
-@click.option("--bucket", default=None, help="Override bucket name.")
-@click.option("--access-key", default=None, help="Override access key.")
-@click.option("--secret-key", default=None, help="Override secret key.")
-@click.option("--alias", default=None, help="Override MinIO/S3 alias name.")
-@click.option("--full", is_flag=True, default=False, help="Download all document bodies.")
-def remote_announce_sync(
-    target: str | None,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias: str | None,
-    full: bool,
-):
-    """Download current server state to remote/ workspace."""
-    from bootstrap.docs.announcements_remote import AnnouncementRemoteSync
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-
-    endpoint, bucket, access_key, secret_key, alias = __resolve_announce_remote_target(
-        target, endpoint, bucket, access_key, secret_key, alias
-    )
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-    sync = AnnouncementRemoteSync(
-        workspace=workspace,
-        target=target or "minio",
-        endpoint=endpoint,
-        bucket=bucket,
-        access_key=access_key,
-        secret_key=secret_key,
-        alias_name=alias,
-        resource_root=bootstrap.config.CONFIGURATION.data_schema.resource_root,
-    )
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Syncing announcements from remote..."))
-    try:
-        sync.sync(full=full)
-    except RuntimeError as e:
-        click.echo(styled([Fore.RED], f"Error: {e}"))
-        exit(2)
-    click.echo(styled([Fore.GREEN], "  Downloaded catalog and pages"))
-    if full:
-        click.echo(styled([Fore.GREEN], "  Downloaded all document bodies"))
-    click.echo(styled([Fore.GREEN], f"  Remote state saved to: {workspace.remote_dir}"))
-
-
-@remote_announce.command("init")
-@click.option(
-    "--target",
-    type=click.Choice(["minio", "s3"]),
-    default=None,
-    help="Remote target (default: minio).",
-)
-@click.option("--endpoint", default=None, help="Override endpoint URL.")
-@click.option("--bucket", default=None, help="Override bucket name.")
-@click.option("--access-key", default=None, help="Override access key.")
-@click.option("--secret-key", default=None, help="Override secret key.")
-@click.option("--alias", default=None, help="Override MinIO/S3 alias name.")
-@click.option(
-    "--force", is_flag=True, default=False, help="Overwrite if remote already has a catalog."
-)
-def remote_announce_init(
-    target: str | None,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias: str | None,
-    force: bool,
-):
-    """Initialize a new empty announcement workspace on the remote.
-
-    Creates catalog.json and active.json with one empty active page,
-    then uploads them.  Fails if the remote already has content unless
-    --force is given.
-    """
-    from bootstrap.docs.announcements_remote import AnnouncementRemoteSync
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-
-    endpoint, bucket, access_key, secret_key, alias = __resolve_announce_remote_target(
-        target, endpoint, bucket, access_key, secret_key, alias
-    )
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-    sync = AnnouncementRemoteSync(
-        workspace=workspace,
-        target=target or "minio",
-        endpoint=endpoint,
-        bucket=bucket,
-        access_key=access_key,
-        secret_key=secret_key,
-        alias_name=alias,
-        resource_root=bootstrap.config.CONFIGURATION.data_schema.resource_root,
-    )
-
-    action = "Initializing" if not force else "Force-initializing"
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"{action} remote announcements..."))
-    try:
-        sync.init_remote(force=force)
-    except RuntimeError as e:
-        click.echo(styled([Fore.RED], f"Error: {e}"))
-        exit(2)
-    click.echo(styled([Fore.GREEN], "  Created catalog.json and active.json"))
-    click.echo(styled([Fore.GREEN], "  Uploaded to remote"))
-    click.echo(styled([Fore.GREEN], f"  Local mirror at: {workspace.remote_dir}"))
-
-
-@remote_announce.command("add")
-@click.option("--id", "entry_id", required=True, help="Entry ID (matches [a-z0-9][a-z0-9._-]*).")
-@click.option("--zh-title", required=True, help="Chinese title.")
-@click.option("--zh-summary", required=True, help="Chinese summary.")
-@click.option("--zh-body", default=None, help="Chinese body text.")
-@click.option(
-    "--zh-body-file",
-    type=click.Path(exists=True, path_type=Path),
-    default=None,
-    help="Chinese body file.",
-)
-@click.option("--en-title", required=True, help="English title.")
-@click.option("--en-summary", required=True, help="English summary.")
-@click.option("--en-body", default=None, help="English body text.")
-@click.option(
-    "--en-body-file",
-    type=click.Path(exists=True, path_type=Path),
-    default=None,
-    help="English body file.",
-)
-@click.option("--published-at", default=None, help="ISO-8601 timestamp (default: now).")
-@click.option("--tags", default="", help="Comma-separated tags.")
-@click.option("--startup", is_flag=True, default=False, help="Show on startup.")
-@click.option("--channels", default="", help="Comma-separated channels.")
-@click.option("--platforms", default="", help="Comma-separated platforms.")
-@click.option("--min-app-version", default=None, help="Minimum app version (semver).")
-@click.option("--max-app-version", default=None, help="Maximum app version (semver).")
-@click.option("--app-version", default=None, help="App version for version announcement.")
-def remote_announce_add(
-    entry_id: str,
-    zh_title: str,
-    zh_summary: str,
-    zh_body: str | None,
-    zh_body_file: Path | None,
-    en_title: str,
-    en_summary: str,
-    en_body: str | None,
-    en_body_file: Path | None,
-    published_at: str | None,
-    tags: str,
-    startup: bool,
-    channels: str,
-    platforms: str,
-    min_app_version: str | None,
-    max_app_version: str | None,
-    app_version: str | None,
-):
-    """Add a new announcement entry to the staging overlay (active page)."""
-    import re as _re
-
-    from bootstrap.docs.announcements_remote import ACTIVE_KEY
-    from bootstrap.docs.announcements_remote import DOCUMENT_ID_PATTERN
-    from bootstrap.docs.announcements_remote import AnnouncementEntry
-    from bootstrap.docs.announcements_remote import AnnouncementLocalization
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-
-    if not _re.match(DOCUMENT_ID_PATTERN, entry_id):
-        raise click.ClickException(
-            f"Invalid entry ID: {entry_id!r} (must match {DOCUMENT_ID_PATTERN})"
-        )
-
-    if zh_body and zh_body_file:
-        raise click.ClickException("Cannot specify both --zh-body and --zh-body-file")
-    if en_body and en_body_file:
-        raise click.ClickException("Cannot specify both --en-body and --en-body-file")
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-    workspace.ensure_remote_directories()
-
-    # Check for duplicate entry ID across remote active page + overlay
-    effective_ids = workspace.get_effective_entry_ids(ACTIVE_KEY)
-    if entry_id in effective_ids:
-        raise click.ClickException(f"Entry {entry_id!r} already exists. Use 'edit' instead.")
-
-    zh_body_text = (
-        zh_body if zh_body else (zh_body_file.read_text(encoding="utf-8") if zh_body_file else "")
-    )
-    en_body_text = (
-        en_body if en_body else (en_body_file.read_text(encoding="utf-8") if en_body_file else "")
-    )
-
-    if not zh_body_text:
-        raise click.ClickException("Chinese body is required (provide --zh-body or --zh-body-file)")
-    if not en_body_text:
-        raise click.ClickException("English body is required (provide --en-body or --en-body-file)")
-
-    zh_hash = workspace.store_document(zh_body_text)
-    en_hash = workspace.store_document(en_body_text)
-
-    if published_at is None:
-        from datetime import datetime as _datetime
-
-        published_at = _datetime.now(UTC).isoformat().replace("+00:00", "Z")
-
-    tags_list = [t.strip() for t in tags.split(",") if t.strip()]
-    channels_list = [c.strip() for c in channels.split(",") if c.strip()]
-    platforms_list = [p.strip() for p in platforms.split(",") if p.strip()]
-
-    if not channels_list:
-        bootstrap.config.DeveloperConfiguration.ensure_loaded()
-        channels_list = [bootstrap.config.DEV_CONFIGURATION.remote.channel.value]
-
-    if not platforms_list:
-        platforms_list = ["android", "ios"]
-
-    new_entry = AnnouncementEntry(
-        id=entry_id,
-        published_at=published_at,
-        tags=tags_list,
-        startup=startup,
-        min_app_version=min_app_version,
-        max_app_version=max_app_version,
-        channels=channels_list,
-        platforms=platforms_list,
-        app_version=app_version,
-        localizations={
-            "zh": AnnouncementLocalization(title=zh_title, summary=zh_summary, body_hash=zh_hash),
-            "en": AnnouncementLocalization(title=en_title, summary=en_summary, body_hash=en_hash),
-        },
-    )
-
-    overlay = workspace.overlay_upsert_entry(ACTIVE_KEY, new_entry)
-    workspace.write_overlay(overlay)
-
-    remote_active_uuid = workspace.get_remote_active_uuid() or "(no remote)"
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Added announcement entry: {entry_id}"))
-    click.echo(styled([Fore.GREEN], f"  page:  {remote_active_uuid}"))
-    click.echo(styled([Fore.GREEN], f"  zh title: {zh_title}"))
-    click.echo(styled([Fore.GREEN], f"  en title: {en_title}"))
-
-
-@remote_announce.command("edit")
-@click.option("--id", "entry_id", required=True, help="Entry ID to edit.")
-@click.option(
-    "--page",
-    "page_uuid",
-    default=None,
-    help="Target page UUID (default: active page). Use to edit archived entries.",
-)
-@click.option("--zh-title", default=None, help="Chinese title.")
-@click.option("--zh-summary", default=None, help="Chinese summary.")
-@click.option("--zh-body", default=None, help="Chinese body text.")
-@click.option(
-    "--zh-body-file",
-    type=click.Path(exists=True, path_type=Path),
-    default=None,
-    help="Chinese body file.",
-)
-@click.option("--en-title", default=None, help="English title.")
-@click.option("--en-summary", default=None, help="English summary.")
-@click.option("--en-body", default=None, help="English body text.")
-@click.option(
-    "--en-body-file",
-    type=click.Path(exists=True, path_type=Path),
-    default=None,
-    help="English body file.",
-)
-@click.option("--published-at", default=None, help="ISO-8601 timestamp.")
-@click.option("--tags", default=None, help="Comma-separated tags.")
-@click.option("--startup/--no-startup", default=None, help="Show on startup.")
-@click.option("--channels", default=None, help="Comma-separated channels.")
-@click.option("--platforms", default=None, help="Comma-separated platforms.")
-@click.option("--min-app-version", default=None, help="Minimum app version (semver).")
-@click.option("--max-app-version", default=None, help="Maximum app version (semver).")
-@click.option("--app-version", default=None, help="App version for version announcement.")
-def remote_announce_edit(
-    entry_id: str,
-    page_uuid: str | None,
-    zh_title: str | None,
-    zh_summary: str | None,
-    zh_body: str | None,
-    zh_body_file: Path | None,
-    en_title: str | None,
-    en_summary: str | None,
-    en_body: str | None,
-    en_body_file: Path | None,
-    published_at: str | None,
-    tags: str | None,
-    startup: bool | None,
-    channels: str | None,
-    platforms: str | None,
-    min_app_version: str | None,
-    max_app_version: str | None,
-    app_version: str | None,
-):
-    """Edit an existing announcement entry in the staging overlay.
-
-    Targets the active page by default.  Use --page UUID to edit an
-    archived entry.
-    """
-    from bootstrap.docs.announcements_remote import ACTIVE_KEY
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-
-    if all(
-        v is None
-        for v in [
-            zh_title,
-            zh_summary,
-            zh_body,
-            zh_body_file,
-            en_title,
-            en_summary,
-            en_body,
-            en_body_file,
-            published_at,
-            tags,
-            startup,
-            channels,
-            platforms,
-            min_app_version,
-            max_app_version,
-            app_version,
-        ]
-    ):
-        raise click.ClickException("No changes specified (provide at least one flag)")
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-    workspace.ensure_remote_directories()
-
-    if zh_body and zh_body_file:
-        raise click.ClickException("Cannot specify both --zh-body and --zh-body-file")
-    if en_body and en_body_file:
-        raise click.ClickException("Cannot specify both --en-body and --en-body-file")
-
-    page_key = page_uuid if page_uuid else ACTIVE_KEY
-
-    # Resolve the current entry content: look in overlay first, then remote
-    overlay = workspace.read_overlay()
-    page_overlay = overlay.pages.get(page_key, {})
-    if entry_id in page_overlay and page_overlay[entry_id] is not None:
-        entry_to_edit = page_overlay[entry_id]
-    elif page_key == ACTIVE_KEY:
-        try:
-            remote_page = workspace.get_remote_page()
-        except FileNotFoundError:
-            raise click.ClickException(
-                "No remote state. Run 'sync' first or use 'add' to create entries."
-            ) from None
-        entry_to_edit = next((e for e in remote_page.entries if e.id == entry_id), None)
-    else:
-        try:
-            remote_page = workspace._read_page(workspace.remote_dir, page_key)
-        except FileNotFoundError:
-            raise click.ClickException(f"Page {page_key!r} not found on remote.") from None
-        entry_to_edit = next((e for e in remote_page.entries if e.id == entry_id), None)
-
-    if entry_to_edit is None:
-        raise click.ClickException(f"Entry {entry_id!r} not found. Use 'add' first.")
-
-    if zh_body:
-        zh_hash = workspace.store_document(zh_body)
-        entry_to_edit.localizations["zh"].body_hash = zh_hash
-    elif zh_body_file:
-        zh_body_text = zh_body_file.read_text(encoding="utf-8")
-        zh_hash = workspace.store_document(zh_body_text)
-        entry_to_edit.localizations["zh"].body_hash = zh_hash
-
-    if en_body:
-        en_hash = workspace.store_document(en_body)
-        entry_to_edit.localizations["en"].body_hash = en_hash
-    elif en_body_file:
-        en_body_text = en_body_file.read_text(encoding="utf-8")
-        en_hash = workspace.store_document(en_body_text)
-        entry_to_edit.localizations["en"].body_hash = en_hash
-
-    if zh_title:
-        entry_to_edit.localizations["zh"].title = zh_title
-    if zh_summary:
-        entry_to_edit.localizations["zh"].summary = zh_summary
-    if en_title:
-        entry_to_edit.localizations["en"].title = en_title
-    if en_summary:
-        entry_to_edit.localizations["en"].summary = en_summary
-    if published_at:
-        entry_to_edit.published_at = published_at
-    if tags is not None:
-        entry_to_edit.tags = [t.strip() for t in tags.split(",") if t.strip()]
-    if startup is not None:
-        entry_to_edit.startup = startup
-    if channels is not None:
-        entry_to_edit.channels = [c.strip() for c in channels.split(",") if c.strip()]
-    if platforms is not None:
-        entry_to_edit.platforms = [p.strip() for p in platforms.split(",") if p.strip()]
-    if min_app_version is not None:
-        entry_to_edit.min_app_version = min_app_version
-    if max_app_version is not None:
-        entry_to_edit.max_app_version = max_app_version
-    if app_version is not None:
-        entry_to_edit.app_version = app_version
-
-    overlay = workspace.overlay_upsert_entry(page_key, entry_to_edit)
-    workspace.write_overlay(overlay)
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Edited announcement entry: {entry_id}"))
-
-
-@remote_announce.command("remove")
-@click.option("--id", "entry_id", required=True, help="Entry ID to remove.")
-def remote_announce_remove(entry_id: str):
-    """Remove an announcement entry from the staging overlay (active page only)."""
-    from bootstrap.docs.announcements_remote import ACTIVE_KEY
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-
-    effective_ids = workspace.get_effective_entry_ids(ACTIVE_KEY)
-    if entry_id not in effective_ids:
-        raise click.ClickException(f"Entry {entry_id!r} not found in active page.")
-
-    overlay = workspace.overlay_remove_entry(entry_id)
-    workspace.write_overlay(overlay)
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Removed announcement entry: {entry_id}"))
-
-
-@remote_announce.command("status")
-@click.option(
-    "--json", "as_json", is_flag=True, default=False, help="Output as machine-readable JSON."
-)
-def remote_announce_status(as_json: bool):
-    """Show effective diff between remote and pending overlay changes.
-
-    Constructs the effective workspace in a temp directory by applying the
-    staging overlay to remote, then diffs against remote.
-    """
-    import tempfile
-
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-    from bootstrap.docs.announcements_remote import compute_status_diff
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-
-    if not (workspace.remote_dir / "catalog.json").exists():
-        click.echo(styled([Fore.YELLOW], "No remote state — run 'sync' first."))
-        exit(1)
-
-    overlay = workspace.read_overlay()
-    has_changes = any(v for v in overlay.pages.values())
-    if not has_changes:
-        click.echo(styled([Fore.GREEN], "No pending changes — overlay is empty."))
-        exit(0)
-
-    temp_dir = Path(tempfile.mkdtemp(prefix="efa-anno-status-"))
-    try:
-        workspace.build_publish_workspace(temp_dir)
-        diff = compute_status_diff(workspace.remote_dir, temp_dir)
-
-        if as_json:
-            click.echo(json.dumps(diff, indent=2))
-            return
-
-        summary = diff["summary"]
-        if summary["added"] == 0 and summary["removed"] == 0 and summary["modified"] == 0:
-            click.echo(styled([Fore.GREEN], "No differences — staging is clean."))
-            exit(0)
-
-        pages = diff.get("pages", {})
-        for page_uuid, page_diff in pages.items():
-            if (
-                page_diff["summary"]["added"] == 0
-                and page_diff["summary"]["removed"] == 0
-                and page_diff["summary"]["modified"] == 0
-            ):
-                continue
-
-            click.echo(styled([Style.BRIGHT], f"Page {page_uuid[:8]}… REMOTE → EFFECTIVE"))
-            click.echo()
-
-            for item in page_diff["added"]:
-                click.echo(
-                    styled([Fore.GREEN], "  +  added     ")
-                    + styled([Style.BRIGHT], f"{item['id']:20}")
-                    + styled(Fore.CYAN, f'zh:"{item["zhTitle"]}"  ')
-                    + styled(Fore.CYAN, f'en:"{item["enTitle"]}"')
-                )
-
-            for item in page_diff["modified"]:
-                change_desc = ", ".join(
-                    f"{k}: {v['from']} → {v['to']}" for k, v in item["changes"].items()
-                )
-                click.echo(
-                    styled([Fore.YELLOW], "  ~  modified  ")
-                    + styled([Style.BRIGHT], f"{item['id']:20}")
-                    + styled(Fore.CYAN, f'zh:"{item["zhTitle"]}"  ')
-                    + styled(Fore.CYAN, f'en:"{item["enTitle"]}"')
-                    + styled(Style.DIM, f"  ({change_desc})")
-                )
-
-            for item in page_diff["removed"]:
-                click.echo(
-                    styled([Fore.RED], "  -  removed   ")
-                    + styled([Style.BRIGHT], f"{item['id']:20}")
-                    + styled(Fore.CYAN, f'zh:"{item["zhTitle"]}"  ')
-                    + styled(Fore.CYAN, f'en:"{item["enTitle"]}"')
-                )
-
-            click.echo()
-
-        click.echo(
-            styled([Style.BRIGHT], "Summary: ")
-            + styled([Fore.GREEN], f"{summary['added']} added, ")
-            + styled([Fore.RED], f"{summary['removed']} removed, ")
-            + styled([Fore.YELLOW], f"{summary['modified']} modified")
-        )
-        click.echo(
-            styled(
-                Style.DIM,
-                f"  Remote: {summary['totalRemote']} entries, "
-                f"Effective: {summary['totalStaging']} entries",
-            )
-        )
-        exit(1)
-    finally:
-        import shutil as _shutil
-
-        _shutil.rmtree(temp_dir, ignore_errors=True)
-
-
-@remote_announce.command("publish")
-@click.option(
-    "--target",
-    type=click.Choice(["minio", "s3"]),
-    default=None,
-    help="Remote target (default: minio).",
-)
-@click.option("--endpoint", default=None, help="Override endpoint URL.")
-@click.option("--bucket", default=None, help="Override bucket name.")
-@click.option("--access-key", default=None, help="Override access key.")
-@click.option("--secret-key", default=None, help="Override secret key.")
-@click.option("--alias", default=None, help="Override MinIO/S3 alias name.")
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    default=False,
-    help="Print files to upload without uploading.",
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    default=False,
-    help="Skip sync and preflight, publish overlay directly.",
-)
-@click.option(
-    "--no-check-remote",
-    is_flag=True,
-    default=False,
-    help="Skip remote compatibility check during preflight.",
-)
-def remote_announce_publish(
-    target: str | None,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias: str | None,
-    dry_run: bool,
-    force: bool,
-    no_check_remote: bool,
-):
-    """Publish staging overlay to remote storage.
-
-    Syncs latest remote, constructs the full workspace in a temp directory
-    (applying the overlay + handling rotation), validates, then uploads.
-    On success, the overlay is cleared and the local remote mirror is updated.
-
-    Use --force to skip sync and preflight (publish overlay directly).
-    """
-    import tempfile
-
-    from bootstrap.docs.announcements_remote import AnnouncementRemoteSync
-    from bootstrap.docs.announcements_remote import AnnouncementWorkspace
-    from bootstrap.docs.announcements_remote import run_preflight_validation
-
-    endpoint, bucket, access_key, secret_key, alias = __resolve_announce_remote_target(
-        target, endpoint, bucket, access_key, secret_key, alias
-    )
-
-    workspace_root = __get_announce_workspace()
-    workspace = AnnouncementWorkspace(workspace_root)
-
-    overlay = workspace.read_overlay()
-    has_changes = any(v for v in overlay.pages.values())
-    if not has_changes:
-        raise click.ClickException("No pending changes — overlay is empty. Run 'add' first.")
-
-    sync = AnnouncementRemoteSync(
-        workspace=workspace,
-        target=target or "minio",
-        endpoint=endpoint,
-        bucket=bucket,
-        access_key=access_key,
-        secret_key=secret_key,
-        alias_name=alias,
-        resource_root=bootstrap.config.CONFIGURATION.data_schema.resource_root,
-    )
-
-    if not force:
-        # Sync the latest remote state
-        try:
-            click.echo(styled([Style.BRIGHT, Fore.CYAN], "Syncing latest remote state..."))
-            sync.sync(full=False)
-            click.echo(styled([Fore.GREEN], "  Remote state synced"))
-        except Exception as e:
-            click.echo(styled([Fore.RED], f"  Sync failed: {e}"))
-            exit(1)
-
-        # Warn if remote active UUID changed (someone else rotated)
-        old_active = None
-        for page_key in overlay.pages:
-            if (
-                page_key != "active"
-                and not (workspace.remote_dir / "pages" / f"{page_key}.json").exists()
-            ):
-                old_active = page_key
-                break
-        if old_active:
-            click.echo(
-                styled(
-                    [Fore.YELLOW],
-                    "  Remote active page changed (rotation by others). "
-                    "Re-applying overlay on new base.",
-                )
-            )
-
-    # Build publish workspace in temp dir
-    temp_dir = Path(tempfile.mkdtemp(prefix="efa-anno-publish-"))
-    try:
-        click.echo(styled([Style.BRIGHT, Fore.CYAN], "Building publish workspace..."))
-        workspace.build_publish_workspace(temp_dir)
-
-        # Count entries for reporting
-        temp_catalog = workspace._read_catalog(temp_dir)
-        total_entries = sum(p.count for p in temp_catalog.pages)
-        body_hashes: set[str] = set()
-        for page_meta in temp_catalog.pages:
-            p = workspace._read_any_page(temp_dir, page_meta.uuid)
-            for entry in p.entries:
-                for loc in entry.localizations.values():
-                    body_hashes.add(loc.body_hash)
-
-        # Preflight validation (skipped when --force)
-        if not force:
-            click.echo(styled([Style.BRIGHT, Fore.GREEN], "Running preflight validation..."))
-            errors = run_preflight_validation(
-                workspace_dir=temp_dir,
-                documents_dir=workspace.documents_dir,
-                remote_dir=workspace.remote_dir if not no_check_remote else None,
-                check_remote=not no_check_remote,
-            )
-            if errors:
-                click.echo(styled([Fore.RED], "Preflight validation failed:"))
-                for error in errors:
-                    click.echo(styled([Fore.RED], f"  - {error}"))
-                exit(1)
-            click.echo(styled([Fore.GREEN], "  Preflight validation passed"))
-
-        if dry_run:
-            click.echo(styled([Style.BRIGHT, Fore.CYAN], "[DRY-RUN] Would publish:"))
-            click.echo(f"  Pages: {len(temp_catalog.pages)}")
-            click.echo(f"  Entries: {total_entries}")
-            click.echo(f"  Documents: {len(body_hashes)}")
-            click.echo(f"  Target: {target or 'minio'}")
-            sync.publish_dir(temp_dir, dry_run=True)
-            return
-
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Publishing announcements..."))
-        sync.publish_dir(temp_dir, dry_run=False)
-
-        # On success: clear overlay, update local remote mirror
-        workspace.clear_overlay()
-        import shutil as _shutil2
-
-        if workspace.remote_dir.exists():
-            _shutil2.rmtree(workspace.remote_dir)
-        _shutil2.copytree(temp_dir, workspace.remote_dir)
-
-        click.echo(
-            styled(
-                [Fore.GREEN],
-                f"  Published {total_entries} entries in {len(temp_catalog.pages)} "
-                f"page(s) with {len(body_hashes)} document bodies.",
-            )
-        )
-        click.echo(styled([Fore.GREEN], "  Overlay cleared."))
-    finally:
-        import shutil as _shutil3
-
-        _shutil3.rmtree(temp_dir, ignore_errors=True)
-
-
-# ---------------------------------------------------------------------------
-
-
-def _require_session(store: SessionStore, operation: str) -> Session:
-    """Load the session, raise ClickException if not present."""
-    try:
-        return store.load()
-    except FileNotFoundError:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-
-
-@remote_session.command("init")
-@click.argument("channel")
-@click.option(
-    "--force-overwrite",
-    is_flag=True,
-    default=False,
-    help="Overwrite an existing session.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_init(
-    channel: str,
-    force_overwrite: bool,
-    schema_root: Path | None,
-):
-    """Create a new staging session for generation assembly."""
-    root = _resolve_schema_root(schema_root)
-    resolved_channel = __validate_remote_channel(channel)
-    mgr = SessionManager(root)
-    mgr.ensure_channel(resolved_channel.value)
-    store = SessionStore(root)
-    try:
-        store.init(
-            resolved_channel.value,
-            force_overwrite=force_overwrite,
-        )
-    except SessionExistsError as e:
-        raise click.ClickException(str(e)) from None
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Session initialized on channel ")
-        + resolved_channel.value
-    )
-    click.echo(styled(Style.DIM, f"  Session file: {store.session_path}"))
-
-
-@remote_session.command("status")
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    default=False,
-    help="Machine-readable output.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_status(as_json: bool, schema_root: Path | None):
-    """Show current session summary."""
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-    if not store.exists():
-        if as_json:
-            click.echo(json.dumps({"active": False}))
-        else:
-            click.echo(styled(Style.DIM, "No active session."))
-        return
-    session = store.load()
-    staged = session.staged
-    if as_json:
-        click.echo(
-            json.dumps(
-                {
-                    "active": True,
-                    "channel": session.channel,
-                    "committed": session.committed,
-                    "staged_counts": {
-                        "resources": len(staged.resources),
-                        "releases": len(staged.releases),
-                    },
-                    "file": str(store.session_path),
-                }
-            )
-        )
-    else:
-        committed_label = ""
-        if session.committed:
-            committed_label = styled([Style.BRIGHT, Fore.GREEN], " (committed)")
-        else:
-            committed_label = styled([Style.BRIGHT, Fore.YELLOW], " (uncommitted)")
-        click.echo(f"Session on channel {session.channel}{committed_label}")
-        click.echo(f"  Staged:      R:{len(staged.resources)} L:{len(staged.releases)}")
-        click.echo(styled(Style.DIM, f"  File:        {store.session_path}"))
-        if session.committed:
-            mgr = SessionManager(root)
-            try:
-                head = mgr.get_head(session.channel)
-                gen_hash = head.generation_hash if head.generation_hash else None
-            except Exception:
-                gen_hash = None
-            if gen_hash:
-                click.echo(styled(Style.DIM, f"  Head:        {gen_hash[:16]}..."))
-
-
-@remote_session.command("discard")
-@click.option(
-    "--force",
-    is_flag=True,
-    default=False,
-    help="Discard even if session is committed.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_discard(force: bool, schema_root: Path | None):
-    """Delete the current session."""
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-    try:
-        store.discard(force=force)
-    except SessionManagerInvalidError:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-    except SessionManagerCommittedError:
-        raise click.ClickException("Session is committed. Use --force to discard.") from None
-    click.echo(f"Session discarded: {store.session_path}")
-
-
-# ---- session add / remove ---------------------------------------------------
-
-
-def _resolve_schema_root(schema_root: Path | None) -> Path:
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    if schema_root is not None:
-        return __resolve_dev_path(schema_root)
-    return __resolve_dev_path(bootstrap.config.DEV_CONFIGURATION.paths.schema_dir)
-
-
-def _validate_add_args(
-    snap_type_flag: str | None,
-    source_hash: str | None,
-    source_file: Path | None,
-) -> str:
-    """Validate mutually exclusive add flags. Returns the snap type."""
-    if snap_type_flag is None:
-        raise click.ClickException("Must specify exactly one of --resource, --release.")
-    if source_hash is None and source_file is None:
-        raise click.ClickException("Must specify exactly one of --hash or --file.")
-    if source_hash is not None and source_file is not None:
-        raise click.ClickException("Cannot specify both --hash and --file.")
-    return snap_type_flag
-
-
-_SNAPSHOT_TYPE_DIR: dict[str, str] = {
-    "resource": "resources",
-    "release": "releases",
-}
-
-
-def _check_snapshot_metadata(
-    snap_type: str,
-    snap_dir: Path,
-) -> None:
-    """Verify that metadata.json matches the expected snapshot type.
-
-    Tries parsing with the expected metadata model. On parse failure
-    (ValidationError) or on success with the wrong model having unexpected
-    fields, raises ClickException.
-    """
-    from bootstrap.remote.models import ReleaseSnapshotMetadata
-    from bootstrap.remote.models import ResourceSnapshotMetadata
-    from bootstrap.remote.models import read_json
-
-    metadata_path = snap_dir / "metadata.json"
-    if not metadata_path.is_file():
-        raise click.ClickException(f"Snapshot directory missing metadata.json: {snap_dir}")
-
-    meta_raw = read_json(metadata_path)
-
-    # Try the expected model first; if it parses OK we are done
-    model_for_type = {
-        "resource": ResourceSnapshotMetadata,
-        "release": ReleaseSnapshotMetadata,
-    }
-
-    # Try expected model — success means match
-    try:
-        model_for_type[snap_type].model_validate(meta_raw)
-        return
-    except Exception:
-        pass
-
-    # If expected model fails, check whether any OTHER model succeeds
-    for other_type, other_model in model_for_type.items():
-        if other_type == snap_type:
-            continue
-        try:
-            other_model.model_validate(meta_raw)
-            raise click.ClickException(
-                f"Snapshot metadata at {snap_dir} declares type '{other_type}', not '{snap_type}'."
-            ) from None
-        except Exception:
-            continue
-
-    raise click.ClickException(
-        f"Snapshot metadata at {snap_dir} is not a valid '{snap_type}' metadata JSON."
-    )
-
-
-def _resolve_snapshot_hash_from_prefix(
-    root: Path,
-    snap_type: str,
-    prefix: str,
-) -> str:
-    """Resolve a hash prefix to a full snapshot hash.
-
-    Lists all snapshots of the given type and matches against the prefix.
-    Raises ClickException if no match or multiple matches are found.
-    """
-    from bootstrap.remote.snapshot import SnapshotStore
-
-    if not prefix:
-        raise click.ClickException("Snapshot hash prefix must not be empty")
-
-    snap_store = SnapshotStore(root)
-    list_for_type: dict[str, object] = {
-        "resource": snap_store.list_resource_snapshots,
-        "release": snap_store.list_release_snapshots,
-    }
-    candidates = [h for h in list_for_type[snap_type]() if h.startswith(prefix)]
-
-    if len(candidates) == 0:
-        raise click.ClickException(f"No {snap_type} snapshot found with prefix '{prefix}'")
-    if len(candidates) == 1:
-        return candidates[0]
-
-    raise click.ClickException(
-        f"Multiple {snap_type} snapshots found with prefix '{prefix}':\n"
-        + "\n".join(f"  {c}" for c in candidates)
-    )
-
-
-def _add_snapshot_by_hash(
-    store: SessionStore,
-    root: Path,
-    snap_type: str,
-    hash_value: str,
-) -> None:
-    """Verify snapshot existence + metadata type, then stage."""
-    from bootstrap.remote.paths import release_snapshot_dir
-    from bootstrap.remote.paths import resource_snapshot_dir
-
-    dir_for_type = {
-        "resource": resource_snapshot_dir,
-        "release": release_snapshot_dir,
-    }
-    snap_dir = dir_for_type[snap_type](root, hash_value)
-    if not snap_dir.is_dir():
-        raise click.ClickException(f"Snapshot {hash_value[:16]}... not found at {snap_dir}")
-    _check_snapshot_metadata(snap_type, snap_dir)
-    store.add_snapshot(snap_type, hash_value)  # type: ignore[arg-type]
-
-
-def _add_snapshot_by_file(
-    store: SessionStore,
-    root: Path,
-    snap_type: str,
-    source_file: Path,
-) -> None:
-    """Read a catalog/registry file, compute snapshot, and stage."""
-    import json as _json
-
-    from bootstrap.remote.models import ReleaseSnapshotMetadata
-    from bootstrap.remote.models import ResourceSnapshotMetadata
-    from bootstrap.remote.snapshot import SnapshotStore
-
-    raw = source_file.read_text(encoding="utf-8")
-    try:
-        data = _json.loads(raw)
-    except _json.JSONDecodeError as e:
-        raise click.ClickException(f"Cannot parse {source_file}: {e}") from None
-
-    snap_store = SnapshotStore(root)
-
-    if snap_type == "resource":
-        from bootstrap.remote.models import make_resource_index
-
-        try:
-            metadata = ResourceSnapshotMetadata.model_validate(data["metadata"])
-            entries = data["entries"]
-        except KeyError as e:
-            raise click.ClickException(
-                f"Invalid resource catalog in {source_file}: "
-                f"missing key {e} (expected 'metadata' and 'entries')"
-            ) from None
-        except Exception as e:
-            raise click.ClickException(
-                f"Cannot parse resource metadata in {source_file}: {e}"
-            ) from None
-
-        index_entries: list[tuple[str, str, int]] = []
-        for entry in entries:
-            index_entries.append((entry["resource_id"], entry["content_hash"], int(entry["size"])))
-        index = make_resource_index(index_entries)
-        hash_value = snap_store.create_resource_snapshot(metadata, index)
-
-    elif snap_type == "release":
-        from bootstrap.remote.blob import BlobStore
-        from bootstrap.remote.hash import ident_hash
-        from bootstrap.remote.models import make_release_index
-
-        try:
-            metadata = ReleaseSnapshotMetadata.model_validate(data["metadata"])
-            release = data["release"]
-        except KeyError as e:
-            raise click.ClickException(
-                f"Invalid release registry in {source_file}: "
-                f"missing key {e} (expected 'metadata' and 'release')"
-            ) from None
-        except Exception as e:
-            raise click.ClickException(
-                f"Cannot parse release metadata in {source_file}: {e}"
-            ) from None
-
-        version = release["version"]
-        blob_store = BlobStore(root)
-
-        for platform_key, platform_data in list(release.items()):
-            if platform_key in ("id", "version"):
-                continue
-            if not isinstance(platform_data, dict):
-                continue
-            for variant, value in list(platform_data.items()):
-                if isinstance(value, str) and value.startswith("/"):
-                    file_path = Path(value)
-                    uri = f"release://{version}/{platform_key}/{variant}"
-                    ihash = ident_hash(uri)
-                    chash = blob_store.store_from_path(file_path, ihash)
-                    platform_data[variant] = {"identifier": uri, "content_hash": chash}
-
-        android = release.get("android")
-        android_dict = None
-        if android is not None:
-            android_dict = {
-                "general": android.get("general"),
-                "armv7": android.get("armv7"),
-                "arm64": android.get("arm64"),
-                "x64": android.get("x64"),
-            }
-        index = make_release_index(
-            release_id=release["id"],
-            version=version,
-            android=android_dict,
-        )
-        hash_value = snap_store.create_release_snapshot(metadata, index)
-
-    else:
-        raise click.ClickException(f"Unknown snapshot type: {snap_type}")
-
-    store.add_snapshot(snap_type, hash_value)  # type: ignore[arg-type]
-
-
-@remote_session.command("add")
-@click.option(
-    "--resource",
-    "snap_type_flag",
-    flag_value="resource",
-    help="Stage a resource snapshot.",
-)
-@click.option(
-    "--release",
-    "snap_type_flag",
-    flag_value="release",
-    help="Stage a release snapshot.",
-)
-@click.option(
-    "--hash",
-    "source_hash",
-    default=None,
-    help="Snapshot hash to stage.",
-)
-@click.option(
-    "--file",
-    "source_file",
-    type=click.Path(exists=True, path_type=Path),
-    default=None,
-    help="File to compute snapshot hash from (checkout catalog, registry).",
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    default=False,
-    help="Add to a committed session.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_add(
-    snap_type_flag: str | None,
-    source_hash: str | None,
-    source_file: Path | None,
-    force: bool,
-    schema_root: Path | None,
-):
-    """Stage a snapshot for the next generation commit."""
-    snap_type = _validate_add_args(snap_type_flag, source_hash, source_file)
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-
-    try:
-        _require_session(store, "add")
-    except click.ClickException:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-
-    if not force:
-        store.ensure_editable()
-
-    if source_hash is not None:
-        full_hash = _resolve_snapshot_hash_from_prefix(root, snap_type, source_hash)
-        _add_snapshot_by_hash(store, root, snap_type, full_hash)
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], f"Staged {snap_type} snapshot ")
-            + f"{full_hash[:16]}..."
-        )
-    elif source_file is not None:
-        _add_snapshot_by_file(store, root, snap_type, source_file)
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], f"Staged {snap_type} snapshot ")
-            + f"from {source_file}"
-        )
-
-
-@remote_session.command("remove")
-@click.option(
-    "--resource",
-    "snap_type_flag",
-    flag_value="resource",
-    help="Remove a resource snapshot.",
-)
-@click.option(
-    "--release",
-    "snap_type_flag",
-    flag_value="release",
-    help="Remove a release snapshot.",
-)
-@click.option(
-    "--hash",
-    "source_hash",
-    required=True,
-    help="Snapshot hash to remove.",
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    default=False,
-    help="Remove from a committed session.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_remove(
-    snap_type_flag: str | None,
-    source_hash: str,
-    force: bool,
-    schema_root: Path | None,
-):
-    """Unstage a snapshot."""
-    if snap_type_flag is None:
-        raise click.ClickException("Must specify exactly one of --resource, --release.")
-
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-
-    try:
-        _require_session(store, "remove")
-    except click.ClickException:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-
-    if not force:
-        store.ensure_editable()
-
-    try:
-        store.remove_snapshot(snap_type_flag, source_hash)  # type: ignore[arg-type]
-    except ValueError as e:
-        raise click.ClickException(str(e)) from None
-
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], f"Removed {snap_type_flag} snapshot ")
-        + f"{source_hash[:16]}..."
-    )
-
-
-# ---- session diff ------------------------------------------------------------
-
-
-def _get_snapshot_summary(
-    root: Path,
-    snap_type: str,
-    hash_value: str,
-) -> str:
-    """Return a human-readable metadata summary for a staged snapshot."""
-    from bootstrap.remote.snapshot import SnapshotStore
-
-    snap_store = SnapshotStore(root)
-    try:
-        if snap_type == "resource":
-            meta, _ = snap_store.load_resource_snapshot(hash_value)
-            return f"server_id={meta.server_id}  game_build={meta.game_build}"
-        elif snap_type == "release":
-            meta, _ = snap_store.load_release_snapshot(hash_value)
-            vmin = meta.version_min or "?"
-            vmax = meta.version_max or "?"
-            return f"version_min={vmin}  version_max={vmax}"
-    except Exception:
-        return "(metadata unavailable)"
-    return ""
-
-
-def _compute_diff(root: Path, session: Session) -> dict:
-    """Compare session staged hashes against the current channel head.
-
-    Returns a dict with keys:
-      channel, head, resources, releases.
-    Each snapshot-type key maps to {"added": [...], "removed": [...], "unchanged": [...]}.
-    """
-    from bootstrap.remote.generation import GenerationStore
-    from bootstrap.remote.head import ChannelHeadStore
-
-    head_store = ChannelHeadStore(root)
-    gen_store = GenerationStore(root)
-
-    head_hash: str | None = None
-    head_sets: dict[str, set[str]] = {
-        "resources": set(),
-        "releases": set(),
-    }
-
-    try:
-        head = head_store._safe_get_head(session.channel)
-        if head and head.generation_hash:
-            head_hash = head.generation_hash
-            generation = gen_store.load(head.generation_hash)
-            for entry in generation.resources.entries:
-                head_sets["resources"].add(entry.snapshot_hash)
-            if generation.release_pointer.snapshot_hash:
-                head_sets["releases"].add(generation.release_pointer.snapshot_hash)
-    except Exception:
-        pass
-
-    session_sets = {
-        "resources": set(session.staged.resources),
-        "releases": set(session.staged.releases),
-    }
-
-    diff: dict = {"channel": session.channel, "head": head_hash}
-    for snap_type in ("resources", "releases"):
-        s_set = session_sets[snap_type]
-        h_set = head_sets[snap_type]
-        diff[snap_type] = {
-            "added": sorted(s_set - h_set),
-            "removed": sorted(h_set - s_set),
-            "unchanged": sorted(s_set & h_set),
-        }
-    return diff
-
-
-@remote_session.command("diff")
-@click.option(
-    "--json",
-    "as_json",
-    is_flag=True,
-    default=False,
-    help="Machine-readable diff output.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_diff(as_json: bool, schema_root: Path | None):
-    """Show changes between staged snapshots and the current channel head."""
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-
-    try:
-        session = _require_session(store, "diff")
-    except click.ClickException:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-
-    diff = _compute_diff(root, session)
-
-    if as_json:
-        click.echo(json.dumps(diff, indent=2))
-        return
-
-    head_str = diff["head"][:16] + "..." if diff["head"] else "none"
-    click.echo(
-        styled([Style.BRIGHT], f'Diff: staging → channel "{diff["channel"]}"')
-        + styled(Style.DIM, f" (head: {head_str})")
-    )
-
-    if diff["head"] is None:
-        click.echo(styled(Style.DIM, "  No channel head (uninitialized channel)."))
-        click.echo(styled(Style.DIM, "  All staged snapshots are new."))
-        click.echo()
-
-    type_labels = {
-        "resources": "Resources",
-        "releases": "Releases",
-    }
-    for snap_type in ("resources", "releases"):
-        data = diff[snap_type]
-        if not data["added"] and not data["removed"] and not data["unchanged"]:
-            continue
-        click.echo(f"\n{type_labels[snap_type]}:")
-        for h in data["added"]:
-            summary = _get_snapshot_summary(root, snap_type.rstrip("s"), h)
-            click.echo(
-                styled([Style.BRIGHT, Fore.GREEN], f"  + {h[:16]}...")
-                + styled(Style.DIM, f"  {summary}")
-            )
-        for h in data["removed"]:
-            summary = _get_snapshot_summary(root, snap_type.rstrip("s"), h)
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], f"  - {h[:16]}...")
-                + styled(Style.DIM, f"  {summary}")
-            )
-        for h in data["unchanged"]:
-            summary = _get_snapshot_summary(root, snap_type.rstrip("s"), h)
-            click.echo(styled(Style.DIM, f"  = {h[:16]}...  {summary}"))
-
-
-# ---- session verify ----------------------------------------------------------
-
-
-def _check_staged_resource_blobs(
-    root: Path,
-    hash_value: str,
-    issues: list,
-) -> None:
-    """Verify all blobs referenced by a staged resource snapshot exist."""
-    from bootstrap.remote.hash import content_hash as _content_hash
-    from bootstrap.remote.hash import ident_hash as _ident_hash
-    from bootstrap.remote.models import ResourceIndex
-    from bootstrap.remote.models import read_pb2
-    from bootstrap.remote.paths import blob_path
-    from bootstrap.remote.paths import resource_snapshot_dir
-    from bootstrap.remote.verify import Issue
-
-    proto_path = resource_snapshot_dir(root, hash_value) / "resources.pb2"
-    try:
-        index = read_pb2(proto_path, ResourceIndex)
-    except Exception:
-        issues.append(
-            Issue(
-                entity=hash_value[:12] + "...",
-                entity_type="resource_snapshot",
-                severity="error",
-                message=f"Cannot read ResourceIndex from {proto_path}",
-            )
-        )
-        return
-
-    for entry in index.entries:
-        ihash = _ident_hash(entry.resource_id)
-        bpath = blob_path(root, ihash, entry.content_hash)
-        if not bpath.is_file():
-            issues.append(
-                Issue(
-                    entity=entry.resource_id,
-                    entity_type="blob",
-                    severity="error",
-                    message=f"Missing blob: {bpath}",
-                )
-            )
-            continue
-
-        try:
-            actual_hash = _content_hash(bpath.read_bytes())
-            if actual_hash != entry.content_hash:
-                issues.append(
-                    Issue(
-                        entity=entry.resource_id,
-                        entity_type="blob",
-                        severity="error",
-                        message=(
-                            f"Content hash mismatch: expected"
-                            f" {entry.content_hash[:12]}..."
-                            f", got {actual_hash[:12]}..."
-                        ),
-                    )
-                )
-        except Exception as exc:
-            issues.append(
-                Issue(
-                    entity=entry.resource_id,
-                    entity_type="blob",
-                    severity="error",
-                    message=str(exc),
-                )
-            )
-
-
-def _verify_staged(root: Path, session: Session) -> list:
-    """Validate staged snapshots across all four verification phases.
-
-    Returns a list of Issue objects.
-    """
-    from bootstrap.remote.hash import snapshot_hash as _snapshot_hash
-    from bootstrap.remote.paths import release_snapshot_dir
-    from bootstrap.remote.paths import resource_snapshot_dir
-    from bootstrap.remote.verify import Issue
-
-    issues: list = []
-
-    dir_for_type: dict[str, callable] = {
-        "resource": resource_snapshot_dir,
-        "release": release_snapshot_dir,
-    }
-    proto_names: dict[str, str] = {
-        "resource": "resources.pb2",
-        "release": "releases.pb2",
-    }
-    staged_map: dict[str, list[str]] = {
-        "resource": session.staged.resources,
-        "release": session.staged.releases,
-    }
-
-    # Phase 1: Per-snapshot integrity
-    # Phase 2: Blob integrity (inline for resource snapshots)
-    for snap_type in ("resource", "release"):
-        proto_name = proto_names[snap_type]
-        staged_hashes = staged_map[snap_type]
-        for h in staged_hashes:
-            snap_dir = dir_for_type[snap_type](root, h)
-            if not snap_dir.is_dir():
-                issues.append(
-                    Issue(
-                        entity=h[:12] + "...",
-                        entity_type=f"{snap_type}_snapshot",
-                        severity="error",
-                        message=f"Directory not found: {snap_dir}",
-                    )
-                )
-                continue
-
-            meta_path = snap_dir / "metadata.json"
-            proto_path = snap_dir / proto_name
-
-            if not meta_path.is_file():
-                issues.append(
-                    Issue(
-                        entity=h[:12] + "...",
-                        entity_type=f"{snap_type}_snapshot",
-                        severity="error",
-                        message="Missing metadata.json",
-                    )
-                )
-                continue
-
-            if not proto_path.is_file():
-                issues.append(
-                    Issue(
-                        entity=h[:12] + "...",
-                        entity_type=f"{snap_type}_snapshot",
-                        severity="error",
-                        message=f"Missing {proto_name}",
-                    )
-                )
-                continue
-
-            # Hash integrity
-            try:
-                files = {
-                    "metadata.json": meta_path.read_bytes(),
-                    proto_name: proto_path.read_bytes(),
-                }
-                computed = _snapshot_hash(snap_type, files)
-                if computed != h:
-                    issues.append(
-                        Issue(
-                            entity=h[:12] + "...",
-                            entity_type=f"{snap_type}_snapshot",
-                            severity="error",
-                            message=(
-                                f"Hash mismatch: expected {h[:12]}..., computed {computed[:12]}..."
-                            ),
-                        )
-                    )
-            except Exception as exc:
-                issues.append(
-                    Issue(
-                        entity=h[:12] + "...",
-                        entity_type=f"{snap_type}_snapshot",
-                        severity="error",
-                        message=f"Hash computation failed: {exc}",
-                    )
-                )
-
-        # Phase 2: Blob check for resources only
-        if snap_type == "resource":
-            for h in staged_hashes:
-                # Skip if snapshot already had directory/meta errors
-                snap_dir = dir_for_type[snap_type](root, h)
-                if snap_dir.is_dir() and (snap_dir / proto_name).is_file():
-                    _check_staged_resource_blobs(root, h, issues)
-
-    # Phase 3: Channel check
-    try:
-        from bootstrap.remote.head import ChannelHeadStore
-
-        head_store = ChannelHeadStore(root)
-        registry = head_store.get_registry()
-        if session.channel not in registry.channels:
-            issues.append(
-                Issue(
-                    entity=session.channel,
-                    entity_type="channel",
-                    severity="warning",
-                    message=(
-                        f"Channel {session.channel!r} not in registry "
-                        "(will be auto-created on commit)"
-                    ),
-                )
-            )
-    except Exception as exc:
-        issues.append(
-            Issue(
-                entity=session.channel,
-                entity_type="channel",
-                severity="warning",
-                message=f"Channel check failed: {exc}",
-            )
-        )
-
-    return issues
-
-
-@remote_session.command("verify")
-@click.option(
-    "--repair",
-    is_flag=True,
-    default=False,
-    help="Attempt automatic repairs.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_verify(repair: bool, schema_root: Path | None):
-    """Validate the staged generation integrity."""
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-
-    try:
-        session = _require_session(store, "verify")
-    except click.ClickException:
-        raise click.ClickException(
-            "No active session. Run './x remote session init' first."
-        ) from None
-
-    if repair:
-        verifier = Verifier(root)
-        fixed = verifier.repair()
-        if fixed > 0:
-            click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Repaired {fixed} entity(ies)."))
-        else:
-            click.echo("No entities needed repair.")
-        return
-
-    issues = _verify_staged(root, session)
-    error_count = sum(1 for i in issues if i.severity == "error")
-
-    staged_counts = {
-        "Resources": len(session.staged.resources),
-        "Releases": len(session.staged.releases),
-    }
-
-    if not issues:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Session verification passed."))
-        for label, count in staged_counts.items():
-            click.echo(styled(Style.DIM, f"  {label}: {count} staged, {count} ok"))
-        return
-
-    click.echo()
-    click.echo(styled([Style.BRIGHT, Fore.RED], f"Verification failed ({error_count} error(s)):"))
-    for issue in issues:
-        color = Fore.RED if issue.severity == "error" else Fore.YELLOW
-        click.echo(
-            f"  {issue.entity[:16] if len(issue.entity) > 16 else issue.entity}"
-            + styled(Style.DIM, f"  [{issue.entity_type}]")
-            + styled([Style.BRIGHT, color], f"  {issue.severity}")
-            + styled(Style.DIM, f"  {issue.message}")
-        )
-
-    if error_count:
-        raise SystemExit(1)
-
-
-# ---- commit ----------------------------------------------------------------
-
-
-@remote_session.command("commit")
-@click.option(
-    "--no-push",
-    is_flag=True,
-    default=False,
-    help="Create generation but do not advance channel head.",
-)
-@click.option(
-    "--force",
-    is_flag=True,
-    default=False,
-    help="Skip verification and override committed session.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_session_commit(no_push: bool, force: bool, schema_root: Path | None):
-    """Assemble a generation from staged snapshots and advance the channel head."""
-    from bootstrap.remote.generation import utc_timestamp
-    from bootstrap.remote.models import GenerationMetadata
-    from bootstrap.remote.models import GenerationPointer
-    from bootstrap.remote.models import GenerationResources
-    from bootstrap.remote.models import ServerIndex
-
-    root = _resolve_schema_root(schema_root)
-    store = SessionStore(root)
-    session = _require_session(store, "commit")
-
-    if session.committed and not force:
-        raise click.ClickException("Session is committed. Use --force to override.")
-
-    if not any(
-        [
-            session.staged.resources,
-            session.staged.releases,
-        ]
-    ):
-        raise click.ClickException(
-            "No snapshots staged. Use './x remote session add' to stage snapshots."
-        )
-
-    if not force:
-        issues = _verify_staged(root, session)
-        error_count = sum(1 for i in issues if i.severity == "error")
-        if error_count:
-            click.echo()
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], f"Verification failed ({error_count} error(s)):")
-            )
-            for issue in issues:
-                color = Fore.RED if issue.severity == "error" else Fore.YELLOW
-                click.echo(
-                    f"  {issue.entity[:16] if len(issue.entity) > 16 else issue.entity}"
-                    + styled(Style.DIM, f"  [{issue.entity_type}]")
-                    + styled([Style.BRIGHT, color], f"  {issue.severity}")
-                    + styled(Style.DIM, f"  {issue.message}")
-                )
-            raise click.ClickException(
-                "Verification failed. Use --force to skip or fix issues first."
-            )
-
-    # Assemble generation
-    mgr = SessionManager(root)
-    snap_store = mgr.snap_store
-    head_store = mgr.head_store
-
-    resolved_channel = __validate_remote_channel(session.channel)
-
-    # Build ServerIndex from staged resource snapshot metadata
-    server_index = ServerIndex()
-    server_index.schema_version = 1
-
-    gen_resources = GenerationResources()
-    gen_resources.schema_version = 1
-
-    server_ids: list[str] = []
-    for hash_val in session.staged.resources:
-        meta, _index = snap_store.load_resource_snapshot(hash_val)
-        entry = server_index.servers.add()
-        entry.server_id = meta.server_id
-        if meta.name:
-            for locale, display_name in meta.name.items():
-                entry.name[locale] = display_name
-        else:
-            entry.name["en"] = meta.server_id
-        entry.game_build = meta.game_build
-        entry.game_version = meta.game_version
-
-        gentry = gen_resources.entries.add()
-        gentry.server_id = meta.server_id
-        gentry.snapshot_hash = hash_val
-
-        server_ids.append(meta.server_id)
-
-    # Build GenerationPointer for releases (last staged hash wins)
-    release_ptr = GenerationPointer()
-    release_ptr.schema_version = 1
-    release_ptr.snapshot_hash = ""
-    release_hash: str | None = None
-    if session.staged.releases:
-        release_hash = session.staged.releases[-1]
-        release_ptr.snapshot_hash = release_hash
-
-    # Determine parent hash from current head
-    current_head = head_store._safe_get_head(resolved_channel)
-    parent = current_head.generation_hash if current_head and current_head.generation_hash else None
-
-    # Create generation metadata
-    ts = utc_timestamp()
-    meta = GenerationMetadata(
-        channel=resolved_channel,
-        timestamp=ts,
-        subject="",
-        parent=parent,
-    )
-
-    mgr.ensure_channel(resolved_channel)
-
-    # Check for idempotent reuse: scan existing generation hashes before creation
-    refs_dir = root / "channels" / "refs"
-    existing_hashes: set[str] = set()
-    if refs_dir.is_dir():
-        for entry in refs_dir.iterdir():
-            if entry.is_dir() and not entry.name.startswith("tmp"):
-                existing_hashes.add(entry.name)
-
-    gen_hash = mgr.create_generation(
-        metadata=meta,
-        server_index=server_index,
-        resources=gen_resources,
-        release_pointer=release_ptr,
-    )
-
-    reused = gen_hash in existing_hashes
-
-    # Advance head (unless --no-push)
-    if not no_push:
-        mgr.push(resolved_channel, gen_hash)
-
-    # Mark session committed
-    store.mark_committed()
-
-    # Output
-    if reused:
-        click.echo(
-            styled(
-                [Style.BRIGHT, Fore.GREEN],
-                f"Generation {gen_hash[:16]}... already exists (reused).",
-            )
-        )
-    else:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Generation created: {gen_hash[:16]}..."))
-
-    if not no_push:
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], f"Head advanced on channel {resolved_channel}")
-        )
-    click.echo(
-        styled(Style.DIM, f"  Parent:        {parent[:16] + '...' if parent else 'none (root)'}")
-    )
-    click.echo(
-        styled(
-            Style.DIM,
-            f"  Resources:     {len(session.staged.resources)} snapshots"
-            f" ({', '.join(server_ids) if server_ids else 'none'})",
-        )
-    )
-    release_label = f"{release_hash[:16]}..." if release_hash else "none"
-    click.echo(styled(Style.DIM, f"  Releases:      {release_label}"))
-
-
-# ---- revert -----------------------------------------------------------------
-
-
-@remote.command("revert")
-@click.argument("channel")
-@click.argument("gen_hash")
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_revert(
-    channel: str,
-    gen_hash: str,
-    schema_root: Path | None,
-):
-    """Revert the channel head to a previous generation (pointer move only)."""
-    root = _resolve_schema_root(schema_root)
-    mgr = SessionManager(root)
-
-    resolved_channel = __validate_remote_channel(channel)
-    mgr.revert(resolved_channel, gen_hash)
-
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], f"Channel {resolved_channel} reverted to")
-        + f" {gen_hash[:16]}..."
-    )
-
-
-# ---- gc ---------------------------------------------------------------------
-
-
-@remote.command("gc")
-@click.option(
-    "--dry-run",
-    is_flag=True,
-    default=False,
-    help="List what would be deleted without actually deleting.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_gc(dry_run: bool, schema_root: Path | None):
-    """Garbage collect unreferenced entities from local V2 storage."""
-    root = _resolve_schema_root(schema_root)
-    mgr = SessionManager(root)
-
-    deleted = mgr.gc(dry_run=dry_run)
-
-    if dry_run:
-        if not deleted:
-            click.echo("Nothing to prune.")
-            return
-        click.echo(f"Would delete {len(deleted)} unreferenced entity(ies):")
-        for path in sorted(deleted):
-            click.echo(f"  {path}")
-    else:
-        if not deleted:
-            click.echo("Nothing to prune.")
-            return
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], f"Pruned {len(deleted)} unreferenced entity(ies):")
-        )
-        for path in sorted(deleted):
-            click.echo(styled(Style.DIM, f"  {path}"))
-
-
-# ---- verify -----------------------------------------------------------------
-
-
-@remote.command("verify")
-@click.option(
-    "--repair",
-    is_flag=True,
-    default=False,
-    help="Attempt to repair issues from workspace origin.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_verify(repair: bool, schema_root: Path | None):
-    """Verify integrity of heads, generations, snapshots, and blobs."""
-    root = _resolve_schema_root(schema_root)
-    mgr = SessionManager(root)
-
-    if repair:
-        fixed = mgr.repair()
-        if fixed > 0:
-            click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Repaired {fixed} entity(ies)."))
-        else:
-            click.echo("No entities needed repair.")
-        return
-
-    all_issues = mgr.verify()
-    total = sum(len(v) for v in all_issues.values())
-    if total == 0:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "All entities passed verification."))
-        return
-
-    for category, issues in all_issues.items():
-        if not issues:
-            continue
-        click.echo(f"\n  {styled([Style.BRIGHT, Fore.CYAN], category.capitalize())}:")
-        for issue in issues:
-            color = Fore.RED if issue.severity == "error" else Fore.YELLOW
-            click.echo(
-                f"    {styled([Style.BRIGHT, color], issue.severity.upper())} "
-                f"[{issue.entity_type}] {issue.entity}: {issue.message}"
-            )
-
-    error_count = sum(sum(1 for i in v if i.severity == "error") for v in all_issues.values())
-    warn_count = total - error_count
-    summary = f"{total} issue(s): {error_count} error(s), {warn_count} warning(s)"
-    if error_count:
-        click.echo(styled([Style.BRIGHT, Fore.RED], f"\nVerification failed — {summary}"))
-    else:
-        click.echo(styled([Style.BRIGHT, Fore.YELLOW], f"\nVerification complete — {summary}"))
-
-
-# ---- publish ----------------------------------------------------------------
-
-
-@remote.command("publish")
-@click.argument("channel")
-@click.option(
-    "--target",
-    type=click.Choice(["minio", "s3"], case_sensitive=False),
-    required=True,
-    help="S3-compatible upload target (reads defaults from efa.dev.toml).",
-)
-@click.option("--endpoint", default=None, help="Override S3-compatible endpoint URL.")
-@click.option("--bucket", default=None, help="Override bucket name.")
-@click.option("--access-key", default=None, help="Override access key.")
-@click.option("--secret-key", default=None, help="Override secret key.")
-@click.option("--alias", "alias_name", default=None, help="Override mc alias name.")
-@click.option(
-    "--workers",
-    type=int,
-    default=8,
-    show_default=True,
-    help="Number of parallel upload workers.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-def remote_publish(
-    channel: str,
-    target: str,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias_name: str | None,
-    workers: int,
-    schema_root: Path | None,
-):
-    """Publish the channel's current head to a remote S3/MinIO bucket."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-
-    if target == "minio":
-        minio_cfg = remote_cfg.require_minio("publish")
-        resolved_endpoint = endpoint or f"http://{remote_cfg.host}:{minio_cfg.port}"
-        resolved_bucket = bucket or minio_cfg.bucket
-        resolved_access_key = access_key or minio_cfg.access_key
-        resolved_secret_key = secret_key or minio_cfg.secret_key
-        resolved_alias = alias_name or minio_cfg.alias
-    else:
-        s3_cfg = remote_cfg.require_s3("publish")
-        resolved_endpoint = endpoint or s3_cfg.endpoint
-        resolved_bucket = bucket or s3_cfg.bucket
-        resolved_access_key = access_key or s3_cfg.access_key
-        resolved_secret_key = secret_key or s3_cfg.secret_key
-        resolved_alias = alias_name or s3_cfg.alias
-
-    root = _resolve_schema_root(schema_root)
-    mgr = SessionManager(root)
-
-    resolved_channel = __validate_remote_channel(channel)
-
-    pub = mgr.make_publisher(
-        endpoint=resolved_endpoint,
-        bucket=resolved_bucket,
-        access_key=resolved_access_key,
-        secret_key=resolved_secret_key,
-        alias_name=resolved_alias,
-        workers=workers,
-    )
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Publishing channel {resolved_channel}..."))
-    pub.publish_all_for_head(resolved_channel)
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Publish complete."))
-
-
-# ---- sync -------------------------------------------------------------------
-
-
-@remote.command("sync")
-@click.option(
-    "--target",
-    type=click.Choice(["minio", "s3"], case_sensitive=False),
-    required=True,
-    help="S3-compatible sync target (reads defaults from efa.dev.toml).",
-)
-@click.option("--endpoint", default=None, help="Override S3-compatible endpoint URL.")
-@click.option("--bucket", default=None, help="Override bucket name.")
-@click.option("--access-key", default=None, help="Override access key.")
-@click.option("--secret-key", default=None, help="Override secret key.")
-@click.option("--alias", "alias_name", default=None, help="Override mc alias name.")
-@click.option(
-    "--depth",
-    type=int,
-    default=-1,
-    show_default=True,
-    help="Max generations to walk (-1 = all).",
-)
-@click.option(
-    "--workers",
-    type=int,
-    default=8,
-    show_default=True,
-    help="Number of parallel download workers.",
-)
-@click.option(
-    "--schema-root",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Schema V2 storage root (default from dev config).",
-)
-@click.option(
-    "--channel",
-    default=None,
-    help="Sync a specific channel (default: all channels).",
-)
-def remote_sync(
-    target: str,
-    endpoint: str | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias_name: str | None,
-    depth: int,
-    workers: int,
-    schema_root: Path | None,
-    channel: str | None,
-):
-    """Sync remote metadata/catalog to local schema root (no blobs).
-
-    By default syncs all channels found in the remote registry.
-    Use --channel to limit to a specific channel.
-    """
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-
-    if target == "minio":
-        minio_cfg = remote_cfg.require_minio("sync")
-        resolved_endpoint = endpoint or f"http://{remote_cfg.host}:{minio_cfg.port}"
-        resolved_bucket = bucket or minio_cfg.bucket
-        resolved_access_key = access_key or minio_cfg.access_key
-        resolved_secret_key = secret_key or minio_cfg.secret_key
-        resolved_alias = alias_name or minio_cfg.alias
-    else:
-        s3_cfg = remote_cfg.require_s3("sync")
-        resolved_endpoint = endpoint or s3_cfg.endpoint
-        resolved_bucket = bucket or s3_cfg.bucket
-        resolved_access_key = access_key or s3_cfg.access_key
-        resolved_secret_key = secret_key or s3_cfg.secret_key
-        resolved_alias = alias_name or s3_cfg.alias
-
-    root = _resolve_schema_root(schema_root)
-    mgr = SessionManager(root)
-
-    syncer = mgr.make_syncer(
-        endpoint=resolved_endpoint,
-        bucket=resolved_bucket,
-        access_key=resolved_access_key,
-        secret_key=resolved_secret_key,
-        alias_name=resolved_alias,
-        workers=workers,
-    )
-
-    if channel is not None:
-        resolved_channel = __validate_remote_channel(channel)
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Syncing channel {resolved_channel}..."))
-        result = syncer.sync_channel(resolved_channel, max_depth=depth)
-        _print_sync_result(result)
-    else:
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Syncing all channels..."))
-        results = syncer.sync_all_channels(max_depth=depth)
-        if not results:
-            click.echo(styled([Style.BRIGHT, Fore.YELLOW], "No channels found in remote registry."))
-            return
-        for _ch, r in results.items():
-            _print_sync_result(r)
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Sync complete."))
-
-
-def _print_sync_result(result: SyncResult) -> None:
-    """Print a human-readable sync result summary."""
-    parts: list[str] = [f"  Channel: {result.channel}"]
-    if result.registry:
-        parts.append("registry ✓")
-    if result.head_meta:
-        parts.append("head ✓")
-    if result.reflog:
-        parts.append("reflog ✓")
-    parts.append(f"generations: {result.generations}")
-    parts.append(f"resource snapshots: {result.resource_snapshots}")
-    parts.append(f"release snapshots: {result.release_snapshots}")
-    click.echo("  " + "  ".join(parts))
-
-
-@remote.group(cls=ClickAliasedGroup)
-def mock():
-    """Remote mock origin commands."""
-
-
-def __materialize_remote_mock(origin_dir: Path, clean: bool) -> None:
-    source_dir = PROJECT_ROOT / "docs" / "examples" / "remote" / "mock-origin"
-    if not source_dir.exists():
-        raise click.ClickException(f"Missing remote mock fixture directory: {source_dir}")
-
-    if clean and origin_dir.exists():
-        shutil.rmtree(origin_dir)
-
-    origin_dir.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copytree(source_dir, origin_dir, dirs_exist_ok=True)
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Materialized remote mock origin: ") + str(origin_dir)
-    )
-
-
-@mock.command("materialize")
-@click.option("--origin-dir", type=click.Path(path_type=Path), default=None)
-@click.option("--clean", is_flag=True, default=False, help="Remove the origin directory first.")
-def remote_mock_materialize(origin_dir: Path | None, clean: bool):
-    """Copy committed remote mock fixtures into the configured origin directory."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-    resolved_origin_dir = __resolve_dev_path(origin_dir or remote_cfg.mock_origin_dir)
-    __materialize_remote_mock(resolved_origin_dir, clean)
-
-
-def __start_minio_remote_mock(
-    *,
-    host: str,
-    port: int,
-    console_port: int,
-    origin_dir: Path,
-    data_dir: Path,
-    bucket: str,
-    access_key: str,
-    secret_key: str,
-    alias_name: str,
-    resource_root: str,
-    channel: Channel,
-    clean_bucket: bool,
-    public_download: bool,
-) -> None:
-    data_dir.mkdir(parents=True, exist_ok=True)
-    endpoint = f"http://{host}:{port}"
-    console_endpoint = f"http://{host}:{console_port}"
-    command = [
-        "minio",
-        "server",
-        str(data_dir),
-        "--address",
-        f"{host}:{port}",
-        "--console-address",
-        f"{host}:{console_port}",
-    ]
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + " ".join(command))
-    if DRY_RUN:
-        return
-
-    minio = get_command("minio")
-    command[0] = minio
-    env = os.environ.copy()
-    env["MINIO_ROOT_USER"] = access_key
-    env["MINIO_ROOT_PASSWORD"] = secret_key
-    process = subprocess.Popen(command, env=env, text=True)
-    try:
-        __wait_for_http(f"{endpoint}/minio/health/ready")
-        if clean_bucket:
-            mc = get_command("mc")
-            bucket_target = f"{alias_name}/{bucket}"
-            redacted = "<redacted>"
-            __execute_command_redacted(
-                [mc, "alias", "set", alias_name, endpoint, access_key, secret_key, "--api", "s3v4"],
-                [mc, "alias", "set", alias_name, endpoint, redacted, redacted, "--api", "s3v4"],
-                "REMOTE PUBLISH ALIAS",
-            )
-            __execute_command_redacted(
-                [mc, "mb", "--ignore-existing", bucket_target],
-                [mc, "mb", "--ignore-existing", bucket_target],
-                "REMOTE CLEAN BUCKET (CREATE)",
-            )
-            __execute_command_redacted(
-                [mc, "rm", "--recursive", "--force", bucket_target],
-                [mc, "rm", "--recursive", "--force", bucket_target],
-                "REMOTE CLEAN BUCKET",
-            )
-        from bootstrap.remote.generation import utc_timestamp
-
-        mock_generation = utc_timestamp().replace("-", "").replace(":", "") + "Z"
-        __publish_remote_origin_to_s3(
-            source_dir=origin_dir,
-            endpoint=endpoint,
-            bucket=bucket,
-            access_key=access_key,
-            secret_key=secret_key,
-            alias_name=alias_name,
-            resource_root=resource_root,
-            channel=channel,
-            generation=mock_generation,
-            public_download=public_download,
-        )
-
-        # Attempt V2 publish if mock origin has V2-structured data
-        v2_heads_dir = origin_dir / resource_root / "channels" / "heads"
-        if v2_heads_dir.is_dir():
-            from bootstrap.remote import Publisher as V2Publisher
-
-            pub = V2Publisher(
-                local_root=origin_dir / resource_root,
-                endpoint=endpoint,
-                bucket=bucket,
-                access_key=access_key,
-                secret_key=secret_key,
-                alias_name=alias_name,
-            )
-            pub.publish_all_for_head(channel.value)
-            click.echo(styled(Style.DIM, "  V2 head published."))
-        else:
-            click.echo(styled(Style.DIM, "  No V2 data in mock origin; skipping V2 publish."))
-
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "MinIO console: ") + console_endpoint)
-        __run_foreground(process, "\nMinIO remote mock interrupted by user.")
-    except Exception:
-        process.terminate()
-        try:
-            process.wait(timeout=10)
-        except subprocess.TimeoutExpired:
-            process.kill()
-            process.wait()
-        raise
-
-
-@mock.command("launch")
-@click.option("--host", default=None, help="Override remote mock host.")
-@click.option("--port", type=int, default=None, help="Override MinIO API port.")
-@click.option("--console-port", type=int, default=None, help="Override MinIO console port.")
-@click.option("--bucket", default=None, help="Override MinIO bucket name.")
-@click.option("--access-key", default=None, help="Override MinIO access key.")
-@click.option("--secret-key", default=None, help="Override MinIO secret key.")
-@click.option("--alias", "alias_name", default=None, help="Override mc alias name.")
-@click.option("--origin-dir", type=click.Path(path_type=Path), default=None)
-@click.option("--data-dir", type=click.Path(path_type=Path), default=None)
-@click.option("--resource-root", default=None, help="Override remote resource root.")
-@click.option("--channel", default=None, help="Override remote channel.")
-@click.option(
-    "--no-materialize", is_flag=True, default=False, help="Do not refresh origin fixtures."
-)
-@click.option(
-    "--clean-origin", is_flag=True, default=False, help="Clean origin before materializing."
-)
-@click.option(
-    "--clean-bucket", is_flag=True, default=False, help="Clean MinIO bucket before mirroring."
-)
-@click.option(
-    "--public-download/--private",
-    "public_download",
-    default=None,
-    help="Configure anonymous bucket downloads.",
-)
-def remote_mock_launch(
-    host: str | None,
-    port: int | None,
-    console_port: int | None,
-    bucket: str | None,
-    access_key: str | None,
-    secret_key: str | None,
-    alias_name: str | None,
-    origin_dir: Path | None,
-    data_dir: Path | None,
-    resource_root: str | None,
-    channel: str | None,
-    no_materialize: bool,
-    clean_origin: bool,
-    clean_bucket: bool,
-    public_download: bool | None,
-):
-    """Launch a local MinIO remote mock."""
-    bootstrap.config.DeveloperConfiguration.ensure_loaded()
-    bootstrap.config.ProjectConfiguration.ensure_loaded()
-    remote_cfg = bootstrap.config.DEV_CONFIGURATION.remote
-    minio = remote_cfg.require_minio("mock")
-
-    resolved_host = host or remote_cfg.host
-    resolved_resource_root = (
-        resource_root or bootstrap.config.CONFIGURATION.data_schema.resource_root
-    )
-    resolved_channel = __validate_remote_channel(channel or remote_cfg.channel.value)
-    resolved_origin_dir = __resolve_dev_path(origin_dir or remote_cfg.mock_origin_dir)
-
-    if not no_materialize:
-        __materialize_remote_mock(resolved_origin_dir, clean_origin)
-    elif not resolved_origin_dir.exists():
-        raise click.ClickException(f"Remote mock origin does not exist: {resolved_origin_dir}")
-
-    resolved_port = port or minio.port
-    resolved_console_port = console_port or minio.console_port
-    resolved_bucket = bucket or minio.bucket
-    resolved_access_key = access_key or minio.access_key
-    resolved_secret_key = secret_key or minio.secret_key
-    resolved_data_dir = __resolve_dev_path(data_dir or minio.data_dir)
-    resolved_alias = alias_name or minio.alias
-    resolved_public_download = (
-        public_download if public_download is not None else minio.public_download
-    )
-
-    origin_url = f"http://{resolved_host}:{resolved_port}/{resolved_bucket}"
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Remote index URL: ")
-        + f"{origin_url.rstrip('/')}/{resolved_resource_root.strip('/')}/channels/{resolved_channel.value}/index.json"
-    )
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "MinIO data path: ") + str(resolved_data_dir))
-    __start_minio_remote_mock(
-        host=resolved_host,
-        port=resolved_port,
-        console_port=resolved_console_port,
-        origin_dir=resolved_origin_dir,
-        data_dir=resolved_data_dir,
-        bucket=resolved_bucket,
-        access_key=resolved_access_key,
-        secret_key=resolved_secret_key,
-        alias_name=resolved_alias,
-        resource_root=resolved_resource_root,
-        channel=resolved_channel,
-        clean_bucket=clean_bucket,
-        public_download=resolved_public_download,
-    )
-
-
-@cli.group(aliases=["env"], cls=ClickAliasedGroup)
-def environment():
-    """Environment related commands. Prefer `x dev env`."""
-
-
-@env.command(
-    "add",
-    context_settings={
-        "ignore_unknown_options": True,
-        "allow_extra_args": True,
-    },
-)
-@click.option(
-    "--python",
-    "--py",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as python packages.\nThis will forward the command to `uv add`.",
-)
-@click.option(
-    "--rust",
-    "--rs",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as rust packages.\nThis will forward the command to `cargo add`.",
-)
-@click.option(
-    "--dart",
-    "--flutter",
-    "--fl",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as dart packages.\nThis will forward the command to `flutter pub add`.",
-)
-@click.option("--dry-run", is_flag=True, default=False, help="Show the command without executing.")
-@click.pass_context
-def dev_env_add(ctx: click.Context, python, rust, dart, dry_run):
-    """Add new tool to the current environment.
-
-    This command accept the following syntax:
-
-    \b
-    If `--python`, `-p` is specified, then all arguments after it are treated as python packages to install.
-        The command will be directly passed to `uv add`.
-    If `--dart`, `-d`, `--flutter`, `-f` is specified, then all arguments after it are treated as dart packages to install.
-        The command will be directly passed to `flutter pub add`.
-    If `--rust`, `-r` is specified, then all arguments after it are treated as rust packages to install.
-        The command will be directly passed to `cargo add`.
-    If none of the above is specified, then:
-        If the package starts with `py:` or `python:`, it is treated as a python package to install.
-            If the package starts with `dev:`, it is treated as a development dependency.
-        If the package starts with `dart:`, `flutter:` or `fl:`, it is treated as a dart package to install.
-            all package literals will be forwarded to `flutter pub add`.
-        If the package starts with `rs:` or `rust:` it is treated as a rust package to install.
-            If the package starts with `dev:`, it is treated as a development dependency.
-            If the package starts with `build:`, it is treated as a build dependency.
-        If not specified, the package won't be installed.
-    """
-
-    if dry_run:
-        click.echo(
-            styled([Style.BRIGHT, Fore.YELLOW], "Warning: ")
-            + "--dry-run is not supported for env add; no package manager command was executed."
-        )
-        return
-
-    if len(list(filter(None, [python, rust, dart]))) > 1:
-        click.echo(
-            styled([Style.BRIGHT, Fore.RED], "Invalid usage: ")
-            + "Only one of --python, --dart/--flutter, --rust can be specified."
-        )
-        exit(1)
-
-    argv = list(ctx.args)
-    if python:
-        if len(argv) == 0:
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], "Invalid usage: ") + "No package specified to add."
-            )
-            exit(1)
-        uv = get_command("uv")
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + f"uv add {' '.join(argv)}"
-        )
-        __execute_command([uv, "add", *argv], "UV ADD OUTPUT")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Python package(s) added successfully."))
-        return
-    elif dart:
-        if len(argv) == 0:
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], "Invalid usage: ") + "No package specified to add."
-            )
-            exit(1)
-        flutter = get_command("flutter")
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
-            + f"flutter pub add {' '.join(argv)}"
-        )
-        __execute_command([flutter, "pub", "add", *argv], "FLUTTER PUB ADD OUTPUT")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Dart package(s) added successfully."))
-        return
-    elif rust:
-        if len(argv) == 0:
-            click.echo(
-                styled([Style.BRIGHT, Fore.RED], "Invalid usage: ") + "No package specified to add."
-            )
-            exit(1)
-        cargo = get_command("cargo")
-        click.echo(
-            styled([Style.BRIGHT, Fore.GREEN], "Executing command: ")
-            + f"cargo add {' '.join(argv)}"
-        )
-        __execute_command([cargo, "add", *argv], "CARGO ADD OUTPUT")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], "Rust package(s) added successfully."))
-        return
-
-    # parse arguments
-    pkgs = {
-        "python": {
-            "norm": [],
-            "dev": [],
-        },
-        "dart": [],
-        "rust": {
-            "norm": [],
-            "dev": [],
-            "build": [],
-        },
-        "unknown": [],
-    }
-    for arg in argv:
-        if arg.startswith(("py:", "python:")):
-            inner = arg.split(":", 1)[1].strip()
-            if len(inner) == 0:
-                pkgs["unknown"].append(arg)
-            elif inner.startswith("dev:"):
-                pkg = inner.split(":", 1)[1].strip()
-                if len(pkg) == 0:
-                    pkgs["unknown"].append(arg)
-                else:
-                    pkgs["python"]["dev"].append(pkg)
-            else:
-                pkgs["python"]["norm"].append(inner)
-        elif arg.startswith(("dart:", "flutter:", "fl:")):
-            inner = arg.split(":", 1)[1].strip()
-            if len(inner) == 0:
-                pkgs["unknown"].append(arg)
-            else:
-                pkgs["dart"].append(inner)
-        elif arg.startswith(("rs:", "rust:")):
-            inner = arg.split(":", 1)[1].strip()
-            if len(inner) == 0:
-                pkgs["unknown"].append(arg)
-            elif inner.startswith("dev:"):
-                pkg = inner.split(":", 1)[1].strip()
-                if len(pkg) == 0:
-                    pkgs["unknown"].append(arg)
-                else:
-                    pkgs["rust"]["dev"].append(pkg)
-            elif inner.startswith("build:"):
-                pkg = inner.split(":", 1)[1].strip()
-                if len(pkg) == 0:
-                    pkgs["unknown"].append(arg)
-                else:
-                    pkgs["rust"]["build"].append(pkg)
-            else:
-                pkgs["rust"]["norm"].append(inner)
-        else:
-            pkgs["unknown"].append(arg)
-
-    if (x := sum(map(len, pkgs["python"].values()))) > 0:
-        click.echo(
-            styled(Fore.GREEN, "Adding ")
-            + styled([Style.BRIGHT, Fore.GREEN], f"{x}")
-            + styled(Fore.GREEN, f" python package{'s' if x > 1 else ''}.")
-        )
-        uv = get_command("uv")
-        norm = pkgs["python"]["norm"]
-        if len(norm) > 0:
-            click.echo(
-                f"  · Adding normal package{'s' if len(norm) > 1 else ''}: " + ", ".join(norm)
-            )
-            __execute_command(
-                [uv, "add", *norm],
-                "UV ADD OUTPUT (NORMAL)",
-            )
-
-        dev = pkgs["python"]["dev"]
-        if len(dev) > 0:
-            click.echo(f"  · Adding dev package{'s' if len(dev) > 1 else ''}: " + ", ".join(dev))
-            __execute_command(
-                [uv, "add", "--dev", *dev],
-                "UV ADD OUTPUT (DEV)",
-            )
-
-    if (x := len(pkgs["dart"])) > 0:
-        click.echo(
-            styled(Fore.GREEN, "Adding ")
-            + styled([Style.BRIGHT, Fore.GREEN], f"{x}")
-            + styled(Fore.GREEN, f" dart package{'s' if x > 1 else ''}.")
-        )
-        flutter = get_command("flutter")
-        click.echo(f"  · Adding package{'s' if x > 1 else ''}: " + ", ".join(pkgs["dart"]))
-        __execute_command(
-            [flutter, "pub", "add", *pkgs["dart"]],
-            "FLUTTER PUB ADD OUTPUT",
-        )
-
-    if (x := sum(map(len, pkgs["rust"].values()))) > 0:
-        click.echo(
-            styled(Fore.GREEN, "Adding ")
-            + styled([Style.BRIGHT, Fore.GREEN], f"{x}")
-            + styled(Fore.GREEN, f" rust package{'s' if x > 1 else ''}.")
-        )
-        cargo = get_command("cargo")
-        norm = pkgs["rust"]["norm"]
-        if len(norm) > 0:
-            click.echo(
-                f"  · Adding normal package{'s' if len(norm) > 1 else ''}: " + ", ".join(norm)
-            )
-            __execute_command(
-                [cargo, "add", *norm],
-                "CARGO ADD OUTPUT (NORMAL)",
-            )
-
-        dev = pkgs["rust"]["dev"]
-        if len(dev) > 0:
-            click.echo(f"  · Adding dev package{'s' if len(dev) > 1 else ''}: " + ", ".join(dev))
-            __execute_command(
-                [cargo, "add", "--dev", *dev],
-                "CARGO ADD OUTPUT (DEV)",
-            )
-
-        build = pkgs["rust"]["build"]
-        if len(build) > 0:
-            click.echo(
-                f"  · Adding build package{'s' if len(build) > 1 else ''}: " + ", ".join(build)
-            )
-            __execute_command(
-                [cargo, "add", "--build", *build],
-                "CARGO ADD OUTPUT (BUILD)",
-            )
-
-    if (x := len(pkgs["unknown"])) > 0:
-        click.echo(
-            styled([Style.BRIGHT, Fore.YELLOW], "Warning: ")
-            + f"{x} unknown argument{'s' if x > 1 else ''} ignored: "
-            + ", ".join(pkgs["unknown"])
-        )
-
-
-@environment.command(
-    context_settings={
-        "ignore_unknown_options": True,
-        "allow_extra_args": True,
-    }
-)
-@click.option(
-    "--python",
-    "--py",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as python packages.\nThis will forward the command to `uv add`.",
-)
-@click.option(
-    "--rust",
-    "--rs",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as rust packages.\nThis will forward the command to `cargo add`.",
-)
-@click.option(
-    "--dart",
-    "--flutter",
-    "--fl",
-    is_flag=True,
-    default=False,
-    help="Treat all arguments as dart packages.\nThis will forward the command to `flutter pub add`.",
-)
-@click.option("--dry-run", is_flag=True, default=False, help="Show the command without executing.")
-@click.pass_context
-def add(ctx: click.Context, python, rust, dart, dry_run):
-    """Add new tool to the current environment. Prefer `x dev env add`."""
-    dev_env_add.callback(python, rust, dart, dry_run)
-
-
-@environment.command()
-def install():
-    """Install all tools in the current environment. Prefer `x dev env install`."""
-    __env_install()
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Environment setup completed successfully."))
-
-
-@environment.command(aliases=["update"])
-def upgrade():
-    """Upgrade all tools in the current environment. Prefer `x dev env upgrade`."""
-    __env_upgrade()
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Environment upgrade completed successfully."))
-
-
-@cli.group(cls=ClickAliasedGroup)
-def build():
-    """Build related commands."""
-
-
-_GENERATOR_TYPES = {"static", "native", "localization", "images"}
-
-
-@build.command("data")
-@click.option(
-    "--skip",
-    "-s",
-    multiple=True,
-    help=f"Skip specified data generators. Values: {', '.join(_GENERATOR_TYPES)}",
-)
-@click.option("--author", default=None, help="Author identifier for the snapshot.")
-@click.option("--description", default=None, help="Description for the snapshot.")
-def data_cmd(skip: list[str], author: str | None, description: str | None):
-    """Build data files."""
-    from bootstrap.data.workspace.generate import run_generator
-
-    to_skip = set()
-    for it in skip:
-        for i in it.split(","):
-            to_skip.add(i.strip().lower())
-
-    if len(x := to_skip.difference(_GENERATOR_TYPES)) > 0:
-        click.echo(
-            styled([Style.BRIGHT, Fore.RED], "Invalid generator type to skip: ") + ", ".join(x)
-        )
-        click.echo("Valid types are: " + ", ".join(_GENERATOR_TYPES))
-        exit(1)
-
-    asyncio.run(
-        run_generator(
-            __get_current_workspace_descriptor(), to_skip, author=author, description=description
-        )
-    )
-
-
-@build.command("announcements", aliases=["anno"])
-def build_announcements_cmd():
-    """Build bundled announcement catalog assets."""
-    from bootstrap.docs import build_bundled_announcements
-
-    try:
-        build_bundled_announcements()
-    except ValueError as exception:
-        raise click.ClickException(str(exception)) from exception
-
-
-_ABI_FLUTTER_TO_APK = {
-    "armeabi-v7a": "armv7",
-    "arm64-v8a": "arm64",
-    "x86_64": "x64",
-}
-
-
-def _build_apk_copy_and_verify(src_apk: Path, src_sha1: Path, dst_apk: Path, dst_sha1: Path):
-    shutil.copy2(src_apk, dst_apk)
-    shutil.copy2(src_sha1, dst_sha1)
-    expected = src_sha1.read_text(encoding="utf-8").strip()
-    actual = get_file_sha1(dst_apk)
-    if expected != actual:
-        raise click.ClickException(
-            f"SHA1 mismatch for {dst_apk.name}: expected {expected}, got {actual}"
-        )
-
-
-@build.command("apk")
-@click.option("--clean", is_flag=True, default=False, help="Run `flutter clean` before building.")
-@click.option("--flavor", default=None, help="Flutter flavor to build (e.g. dev, prod).")
-@click.option(
-    "--debug", is_flag=True, default=False, help="Build debug APK (single ABI only, no split)."
-)
-def build_apk_cmd(clean: bool, flavor: str | None, debug: bool):
-    """Build Android APKs with versioned filenames."""
-    ProjectConfiguration.ensure_loaded()
-    version = bootstrap.config.CONFIGURATION.version
-    ver = version.render_full()
-    output_dir = PROJECT_ROOT / "cache" / "releases" / "apk" / ver
-    output_dir.mkdir(parents=True, exist_ok=True)
-    apk_source = PROJECT_ROOT / "build" / "app" / "outputs" / "flutter-apk"
-
-    flutter = get_command("flutter")
-    flavor_args = [f"--flavor={flavor}"] if flavor else []
-
-    if clean:
-        __execute_command([flutter, "clean"], "CLEANING BUILD ARTIFACTS")
-
-    if debug:
-        __execute_command([flutter, "build", "apk", "--debug", *flavor_args], "BUILDING DEBUG APK")
-        src_prefix = f"app-{flavor}-" if flavor else "app-"
-        src_apk = apk_source / f"{src_prefix}debug.apk"
-        src_sha1 = apk_source / f"{src_prefix}debug.apk.sha1"
-        if not src_apk.exists():
-            raise click.ClickException(f"Expected debug APK not found: {src_apk}")
-        dst_apk = output_dir / f"{ver}-android-debug.apk"
-        dst_sha1 = output_dir / f"{ver}-android-debug.apk.sha1"
-        _build_apk_copy_and_verify(src_apk, src_sha1, dst_apk, dst_sha1)
-    else:
-        __execute_command([flutter, "build", "apk", *flavor_args], "BUILDING GENERAL APK")
-        src_apk = apk_source / (f"app-{flavor}-release.apk" if flavor else "app-release.apk")
-        src_sha1 = apk_source / (
-            f"app-{flavor}-release.apk.sha1" if flavor else "app-release.apk.sha1"
-        )
-        if not src_apk.exists():
-            raise click.ClickException(f"Expected general APK not found: {src_apk}")
-        dst_apk = output_dir / f"{ver}-android.apk"
-        dst_sha1 = output_dir / f"{ver}-android.apk.sha1"
-        _build_apk_copy_and_verify(src_apk, src_sha1, dst_apk, dst_sha1)
-
-        __execute_command(
-            [flutter, "build", "apk", "--split-per-abi", *flavor_args],
-            "BUILDING SPLIT ABI APKS",
-        )
-        for flutter_abi, apk_suffix in _ABI_FLUTTER_TO_APK.items():
-            src_apk = apk_source / (
-                f"app-{flavor}-{flutter_abi}-release.apk"
-                if flavor
-                else f"app-{flutter_abi}-release.apk"
-            )
-            src_sha1 = apk_source / (
-                f"app-{flavor}-{flutter_abi}-release.apk.sha1"
-                if flavor
-                else f"app-{flutter_abi}-release.apk.sha1"
-            )
-            if not src_apk.exists():
-                raise click.ClickException(f"Expected ABI APK not found: {src_apk}")
-            dst_apk = output_dir / f"{ver}-android-{apk_suffix}.apk"
-            dst_sha1 = output_dir / f"{ver}-android-{apk_suffix}.apk.sha1"
-            _build_apk_copy_and_verify(src_apk, src_sha1, dst_apk, dst_sha1)
-
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Build complete. Output: {output_dir}"))
-    for f in sorted(output_dir.iterdir()):
-        if f.is_file():
-            size = get_bin_size(f.stat().st_size)
-            click.echo(f"  {f.name} ({size})")
-
-
-_PLATFORM_DIR = {
-    "android": "apk",
-}
-
-_PLATFORM_SUFFIX = {
-    "android": ".apk",
-}
-
-
-@build.command("release")
-@click.option(
-    "--platform",
-    default=None,
-    type=click.Choice(["android"]),
-    help="Platform to produce a fragment for.",
-)
-@click.option(
-    "--merge",
-    "merge_files",
-    multiple=True,
-    type=click.Path(exists=True, path_type=Path),
-    help="Fragment JSON files to merge (repeatable).",
-)
-@click.option(
-    "--output",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="Output file path (default: auto-determined within cache, or stdout for '-').",
-)
-@click.option("--release-id", default=None, help="Override release.id (default: rel-{version}).")
-@click.option("--version-min", default=None, help="Override minimum version string.")
-@click.option("--version-max", default=None, help="Override maximum version string.")
-def build_release_cmd(
-    platform: str | None,
-    merge_files: list[Path],
-    output: Path | None,
-    release_id: str | None,
-    version_min: str | None,
-    version_max: str | None,
-):
-    """Build or merge release registry fragments."""
-    if platform and merge_files:
-        raise click.ClickException("Cannot use --platform and --merge together.")
-    if not platform and not merge_files:
-        raise click.ClickException("Must specify --platform or --merge.")
-
-    ProjectConfiguration.ensure_loaded()
-    version = bootstrap.config.CONFIGURATION.version
-    ver = version.render_full()
-    ver_semver = version.render_semver()
-
-    if platform:
-        _build_release_platform(
-            platform, ver, ver_semver, output, release_id, version_min, version_max
-        )
-    else:
-        _build_release_merge(list(merge_files), ver, output)
-
-
-def _build_release_platform(
-    platform: str,
-    ver: str,
-    ver_semver: str,
-    output: Path | None,
-    release_id: str | None,
-    version_min: str | None,
-    version_max: str | None,
-) -> None:
-    dir_name = _PLATFORM_DIR[platform]
-    suffix = _PLATFORM_SUFFIX[platform]
-    src_dir = PROJECT_ROOT / "cache" / "releases" / dir_name / ver
-    if not src_dir.is_dir():
-        raise click.ClickException(f"Build directory not found: {src_dir}")
-
-    artifacts: dict[str, str] = {}
-    for f in sorted(src_dir.iterdir()):
-        if f.is_file() and f.suffix == suffix:
-            name = f.name
-            prefix = f"{ver}-{platform}-"
-            if name.startswith(prefix):
-                variant = name[len(prefix) : -len(suffix)]
-                artifacts[variant] = str(f.resolve())
-            elif name == f"{ver}-{platform}{suffix}":
-                artifacts["general"] = str(f.resolve())
-
-    if not artifacts:
-        raise click.ClickException(f"No {suffix} files found in {src_dir}")
-
-    rel_id = release_id or f"rel-{ver}"
-    vmin = version_min or ver_semver
-    vmax = version_max or ver_semver
-
-    data: dict[str, object] = {
-        "metadata": {
-            "versionMin": vmin,
-            "versionMax": vmax,
-            "offerings": [platform],
-            "releaseCount": 1,
-            "createdAt": _utcnow_iso(),
-        },
-        "release": {
-            "id": rel_id,
-            "version": ver,
-            platform: artifacts,
-        },
-    }
-
-    _emit_release_json(data, output, src_dir / f"{ver}-{platform}.json")
-
-
-def _build_release_merge(fragments: list[Path], ver: str, output: Path | None) -> None:
-    merged_metadata: dict = {}
-    merged_release: dict = {}
-    all_paths: list[Path] = []
-
-    for fp in fragments:
-        raw = fp.read_text(encoding="utf-8")
-        try:
-            data = json.loads(raw)
-        except json.JSONDecodeError as e:
-            raise click.ClickException(f"Cannot parse {fp}: {e}") from None
-
-        meta = data.get("metadata", {})
-        rel = data.get("release", {})
-
-        for offering in meta.get("offerings", []):
-            if offering not in merged_metadata.setdefault("offerings", []):
-                merged_metadata["offerings"].append(offering)
-
-        merged_metadata.setdefault("versionMin", meta.get("versionMin"))
-        merged_metadata.setdefault("versionMax", meta.get("versionMax"))
-        merged_metadata.setdefault("createdAt", meta.get("createdAt"))
-
-        for pkey, pdict in rel.items():
-            if pkey in ("id", "version"):
-                merged_release.setdefault(pkey, pdict)
-            elif isinstance(pdict, dict):
-                for _variant, path_str in pdict.items():
-                    if isinstance(path_str, str):
-                        pp = Path(path_str)
-                        all_paths.append(pp)
-
-        for pkey, pdict in rel.items():
-            if pkey not in ("id", "version") and isinstance(pdict, dict):
-                merged_release.setdefault(pkey, pdict)
-
-    missing = [str(p) for p in all_paths if not p.exists()]
-    if missing:
-        raise click.ClickException(
-            "Missing files referenced in fragments:\n  " + "\n  ".join(missing)
-        )
-
-    merged_metadata["releaseCount"] = 1
-    data: dict[str, object] = {
-        "metadata": merged_metadata,
-        "release": merged_release,
-    }
-
-    default_output = PROJECT_ROOT / "cache" / "releases" / "merge" / f"{ver}.json"
-    _emit_release_json(data, output, default_output)
-
-
-def _emit_release_json(data: dict[str, object], output: Path | None, default: Path) -> None:
-    import json as _json
-
-    text = _json.dumps(data, indent=2, ensure_ascii=False, sort_keys=False) + "\n"
-    if output is None:
-        default.parent.mkdir(parents=True, exist_ok=True)
-        default.write_text(text, encoding="utf-8")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Written: {default}"))
-    elif output == Path("-"):
-        click.echo(text, nl=False)
-    else:
-        output.parent.mkdir(parents=True, exist_ok=True)
-        output.write_text(text, encoding="utf-8")
-        click.echo(styled([Style.BRIGHT, Fore.GREEN], f"Written: {output}"))
-
-
-def _utcnow_iso() -> str:
-    from datetime import UTC
-    from datetime import datetime
-
-    return datetime.now(UTC).isoformat(timespec="seconds")
-
-
-@build.command("list", aliases=["ls"])
-@click.option("--apps", is_flag=True, default=False, help="Show APK builds.")
-@click.option("--resources", is_flag=True, default=False, help="Show resource snapshots.")
-@click.option("--releases", is_flag=True, default=False, help="Show release snapshots.")
-@click.option("--generations", is_flag=True, default=False, help="Show generations.")
-def build_list_cmd(apps: bool, resources: bool, releases: bool, generations: bool):
-    """List local build / cache items."""
-    from bootstrap.remote.generation import GenerationStore
-    from bootstrap.remote.head import ChannelHeadStore
-    from bootstrap.remote.snapshot import SnapshotStore
-
-    def _dir_size(dir_path: Path) -> int:
-        if not dir_path.is_dir():
-            return 0
-        total = 0
-        try:
-            for f in dir_path.rglob("*"):
-                if f.is_file():
-                    total += f.stat().st_size
-        except PermissionError:
-            pass
-        return total
-
-    def _scan_apk_versions(apk_root: Path) -> tuple[list[Path], list[Path]]:
-        versions: list[Path] = []
-        files: list[Path] = []
-        if apk_root.is_dir():
-            for vd in sorted(apk_root.iterdir()):
-                if vd.is_dir():
-                    versions.append(vd)
-                    for f in vd.iterdir():
-                        if f.is_file():
-                            files.append(f)
-        return versions, files
-
-    any_opt = apps or resources or releases or generations
-    schema_root = _resolve_schema_root(None)
-
-    if not any_opt:
-        # --- Summary --------------------------------------------------------------
-        click.echo()
-        click.echo(styled([Style.BRIGHT], "Local Build Cache Summary"))
-        click.echo(styled([Style.BRIGHT], "=" * 27))
-
-        # APK builds
-        apk_root = PROJECT_ROOT / "cache" / "releases" / "apk"
-        apk_versions, apk_files = _scan_apk_versions(apk_root)
-        apk_total = sum(f.stat().st_size for f in apk_files)
-        if apk_versions:
-            click.echo(
-                f"  {styled(Fore.GREEN, 'APK builds:')}            "
-                f"{styled([Style.BRIGHT], str(len(apk_versions)))} version(s)"
-                f"  ({get_bin_size(apk_total)})"
-            )
-        else:
-            click.echo(f"  {styled(Fore.GREEN, 'APK builds:')}            (none)")
-
-        # Schema items
-        if schema_root.is_dir():
-            snap_store = SnapshotStore(schema_root)
-            gen_store = GenerationStore(schema_root)
-            head_store = ChannelHeadStore(schema_root)
-
-            res_hashes = snap_store.list_resource_snapshots()
-            rel_hashes = snap_store.list_release_snapshots()
-            all_gens = gen_store.list_all()
-            registry = head_store.get_registry()
-
-            for label, hashes, asset_dir in [
-                ("Resource snapshots:", res_hashes, "resources"),
-                ("Release snapshots:", rel_hashes, "releases"),
-            ]:
-                if hashes:
-                    total = sum(_dir_size(schema_root / "assets" / asset_dir / h) for h in hashes)
-                    click.echo(
-                        f"  {styled(Fore.GREEN, label):30s}"
-                        f"{styled([Style.BRIGHT], str(len(hashes)))} item(s)"
-                        f"  ({get_bin_size(total)})"
-                    )
-                else:
-                    click.echo(f"  {styled(Fore.GREEN, label):30s}(none)")
-
-            if all_gens:
-                gen_total = sum(_dir_size(schema_root / "channels" / "refs" / h) for h in all_gens)
-                click.echo(
-                    f"  {styled(Fore.GREEN, 'Generations:'):30s}"
-                    f"{styled([Style.BRIGHT], str(len(all_gens)))} item(s)"
-                    f"  ({get_bin_size(gen_total)})"
-                )
-            else:
-                click.echo(f"  {styled(Fore.GREEN, 'Generations:'):30s}(none)")
-
-            channels = list(registry.channels.keys())
-            if channels:
-                click.echo(
-                    f"  {styled(Fore.GREEN, 'Channels:'):30s}"
-                    f"{styled([Style.BRIGHT], str(len(channels)))} ({', '.join(channels)})"
-                )
-            else:
-                click.echo(f"  {styled(Fore.GREEN, 'Channels:'):30s}(none)")
-
-            # Blobs
-            blobs_dir = schema_root / "assets" / "blobs"
-            if blobs_dir.is_dir():
-                blobs_count = 0
-                blobs_total = 0
-                for f in blobs_dir.rglob("*"):
-                    if f.is_file():
-                        blobs_count += 1
-                        blobs_total += f.stat().st_size
-                click.echo(
-                    f"  {styled(Fore.GREEN, 'Content blobs:'):30s}"
-                    f"{styled([Style.BRIGHT], str(blobs_count))} file(s)"
-                    f"  ({get_bin_size(blobs_total)})"
-                )
-            else:
-                click.echo(f"  {styled(Fore.GREEN, 'Content blobs:'):30s}(none)")
-        else:
-            click.echo(
-                styled([Style.BRIGHT, Fore.YELLOW], "  Schema cache: ")
-                + styled(Fore.YELLOW, f"not found ({schema_root})")
-            )
-
-        click.echo()
-    else:
-        # --- Detailed listings ---------------------------------------------------
-        first = True
-
-        if apps:
-            if not first:
-                click.echo()
-            first = False
-            apk_root = PROJECT_ROOT / "cache" / "releases" / "apk"
-            apk_versions, apk_files = _scan_apk_versions(apk_root)
-
-            click.echo(styled([Style.BRIGHT, Fore.GREEN], f"\nAPK Builds ({len(apk_versions)})"))
-            click.echo(styled([Style.BRIGHT], "=" * 50))
-            if not apk_versions:
-                click.echo("  (none)")
-            else:
-                for vd in apk_versions:
-                    v_total = sum(f.stat().st_size for f in vd.iterdir() if f.is_file())
-                    click.echo(f"  {styled([Style.BRIGHT], vd.name)}  ({get_bin_size(v_total)})")
-                    for f in sorted(vd.iterdir()):
-                        if f.is_file():
-                            click.echo(f"    {f.name}  ({get_bin_size(f.stat().st_size)})")
-
-        if resources:
-            if not first:
-                click.echo()
-            first = False
-            if not schema_root.is_dir():
-                click.echo(
-                    styled([Style.BRIGHT, Fore.YELLOW], "Schema cache not found: ")
-                    + str(schema_root)
-                )
-            else:
-                snap_store = SnapshotStore(schema_root)
-                res_hashes = snap_store.list_resource_snapshots()
-                click.echo(
-                    styled([Style.BRIGHT, Fore.GREEN], f"\nResource Snapshots ({len(res_hashes)})")
-                )
-                click.echo(styled([Style.BRIGHT], "=" * 80))
-                if not res_hashes:
-                    click.echo("  (none)")
-                else:
-                    click.echo(
-                        f"  {'Hash':9s}  {'Server':13s}  {'Build':13s}"
-                        f"  {'Version':11s}  {'Resources':>10s}  Created"
-                    )
-                    click.echo(
-                        f"  {'-' * 9}  {'-' * 13}  {'-' * 13}  {'-' * 11}  {'-' * 10}  {'-' * 25}"
-                    )
-                    for h in res_hashes:
-                        try:
-                            meta, _ = snap_store.load_resource_snapshot(h)
-                            click.echo(
-                                f"  {h[:9]:9s}  {meta.server_id:13s}  {meta.game_build:13s}"
-                                f"  {meta.game_version:11s}  {meta.resource_count!s:>10s}"
-                                f"  {meta.created_at}"
-                            )
-                        except Exception:
-                            click.echo(f"  {h[:9]:9s}  {styled(Fore.RED, '[ERR]')}")
-
-        if releases:
-            if not first:
-                click.echo()
-            first = False
-            if not schema_root.is_dir():
-                click.echo(
-                    styled([Style.BRIGHT, Fore.YELLOW], "Schema cache not found: ")
-                    + str(schema_root)
-                )
-            else:
-                snap_store = SnapshotStore(schema_root)
-                rel_hashes = snap_store.list_release_snapshots()
-                click.echo(
-                    styled([Style.BRIGHT, Fore.GREEN], f"\nRelease Snapshots ({len(rel_hashes)})")
-                )
-                click.echo(styled([Style.BRIGHT], "=" * 70))
-                if not rel_hashes:
-                    click.echo("  (none)")
-                else:
-                    click.echo(f"  {'Hash':9s}  {'Version Range':21s}  {'Releases':>10s}  Created")
-                    click.echo(f"  {'-' * 9}  {'-' * 21}  {'-' * 10}  {'-' * 25}")
-                    for h in rel_hashes:
-                        try:
-                            meta, _ = snap_store.load_release_snapshot(h)
-                            v_range = f"{meta.version_min or '—'} ~ {meta.version_max or '—'}"
-                            click.echo(
-                                f"  {h[:9]:9s}  {v_range:21s}"
-                                f"  {meta.release_count!s:>10s}  {meta.created_at}"
-                            )
-                        except Exception:
-                            click.echo(f"  {h[:9]:9s}  {styled(Fore.RED, '[ERR]')}")
-
-        if generations:
-            if not first:
-                click.echo()
-            first = False
-            if not schema_root.is_dir():
-                click.echo(
-                    styled([Style.BRIGHT, Fore.YELLOW], "Schema cache not found: ")
-                    + str(schema_root)
-                )
-            else:
-                gen_store = GenerationStore(schema_root)
-                all_gens = gen_store.list_all()
-                click.echo(styled([Style.BRIGHT, Fore.GREEN], f"\nGenerations ({len(all_gens)})"))
-                click.echo(styled([Style.BRIGHT], "=" * 80))
-                if not all_gens:
-                    click.echo("  (none)")
-                else:
-                    click.echo(
-                        f"  {'Hash':9s}  {'Channel':10s}  {'Parent':9s}  {'Subject':20s}  Timestamp"
-                    )
-                    click.echo(f"  {'-' * 9}  {'-' * 10}  {'-' * 9}  {'-' * 20}  {'-' * 25}")
-                    for h, gen in all_gens.items():
-                        parent_short = (
-                            (gen.metadata.parent or "")[:9] if gen.metadata.parent else "—"
-                        )
-                        subject = gen.metadata.subject or ""
-                        click.echo(
-                            f"  {h[:9]:9s}  {gen.metadata.channel:10s}"
-                            f"  {parent_short:9s}  {subject[:19]:20s}"
-                            f"  {gen.metadata.timestamp}"
-                        )
-
-
-@cli.group(aliases=["rel"], cls=ClickAliasedGroup)
-def release():
-    """Pre-release workflow commands."""
-
-
-# --- release version ---
-
-
-@release.group("version", cls=ClickAliasedGroup)
-def release_version():
-    """Version management — show, sync, and bump the canonical version."""
-
-
-@release_version.command("show")
-def release_version_show():
-    """Display the current version from efa.config.toml."""
-    from bootstrap.release.version import load_version
-
-    v = load_version()
-    is_pre = v.is_prerelease()
-
-    lines = [
-        ("Canonical version", v.render_full()),
-        ("  major", str(v.major)),
-        ("  minor", str(v.minor)),
-        ("  patch", str(v.patch)),
-        ("  pre-release", f"{v.pre_label}.{v.pre_num}" if is_pre else "(none)"),
-        ("  build", str(v.build)),
-        ("", ""),
-        ("Full string", v.render_full()),
-        ("Semver only", v.render_semver()),
-        ("Git tag", v.render_tag()),
-    ]
-    label_width = max(len(label) for label, _ in lines if label)
-    for label, value in lines:
-        if not label:
-            click.echo("")
-        else:
-            click.echo(
-                styled([Style.BRIGHT, Fore.CYAN], label.ljust(label_width))
-                + "  "
-                + styled([Style.BRIGHT], value)
-            )
-
-
-@release_version.command("sync")
-@click.option("--dry-run", is_flag=True, default=False, help="Show what would be written.")
-def release_version_sync(dry_run: bool):
-    """Sync version from efa.config.toml to all target files."""
-    from bootstrap.release.version import load_version
-    from bootstrap.release.version import sync_all
-
-    v = load_version()
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Syncing version ")
-        + styled([Style.BRIGHT], v.render_full())
-    )
-
-    report = sync_all(v, dry_run=dry_run)
-    for t in report.synced:
-        click.echo(styled([Fore.GREEN], f"  {t.label:40s} -> {t.expected}"))
-    for err in report.errors:
-        click.echo(styled([Fore.RED], f"  ERROR: {err}"))
-
-    if dry_run:
-        click.echo(styled([Fore.YELLOW], "  (dry-run — no files were modified)"))
-
-
-@release_version.command("bump")
-@click.argument("level", type=click.Choice(["major", "minor", "patch"]), required=False)
-@click.option("--pre-label", default=None, help="Set pre-release label (e.g. 'beta', 'rc').")
-@click.option("--pre-num", type=int, default=None, help="Set pre-release number.")
-@click.option("--build", "build_num", type=int, default=None, help="Set build number.")
-@click.option(
-    "--clear-pre", is_flag=True, default=False, help="Remove pre-release (promote to release)."
-)
-@click.option("--dry-run", is_flag=True, default=False, help="Show what would be done.")
-def release_version_bump(
-    level: str | None,
-    pre_label: str | None,
-    pre_num: int | None,
-    build_num: int | None,
-    clear_pre: bool,
-    dry_run: bool,
-):
-    """Bump the canonical version and sync all target files.
-
-    LEVEL must be one of: major, minor, patch.
-
-    \b
-    Bump patch with pre-release:
-        ./x release version bump patch --pre-label beta --pre-num 1
-
-    \b
-    Bump minor and clear pre-release (final release):
-        ./x release version bump minor --clear-pre
-
-    \b
-    Only update pre-release number:
-        ./x release version bump --pre-label beta --pre-num 5
-
-    \b
-    Only update build number:
-        ./x release version bump --build 42
-    """
-    from bootstrap.release.version import load_version
-    from bootstrap.release.version import sync_all
-    from bootstrap.release.version import write_config_version
-
-    v = load_version()
-    old_ver = v.render_full()
-
-    if level is not None:
-        if level == "major":
-            v.bump_major()
-        elif level == "minor":
-            v.bump_minor()
-        elif level == "patch":
-            v.bump_patch()
-
-    if clear_pre:
-        v.clear_prerelease()
-    else:
-        if pre_label is not None:
-            v.pre_label = pre_label
-            if v.pre_num == 0 and pre_num is None:
-                v.pre_num = 1
-        if pre_num is not None:
-            v.pre_num = pre_num
-            if v.pre_num > 0 and not v.pre_label:
-                raise click.ClickException("pre_label is required when setting pre_num > 0")
-
-    if build_num is not None:
-        v.build = build_num
-
-    new_ver = v.render_full()
-    click.echo(
-        styled([Style.BRIGHT, Fore.CYAN], "Bumping: ")
-        + styled([Style.BRIGHT], f"{old_ver} -> {new_ver}")
-    )
-
-    if dry_run:
-        click.echo(styled([Fore.YELLOW], "  (dry-run — no files were modified)"))
-        return
-
-    write_config_version(v, dry_run=False)
-    click.echo(styled([Fore.GREEN], "  Updated efa.config.toml"))
-
-    report = sync_all(v, dry_run=False)
-    for t in report.synced:
-        click.echo(styled([Fore.GREEN], f"  Synced {t.label} -> {t.expected}"))
-    for err in report.errors:
-        click.echo(styled([Fore.RED], f"  ERROR: {err}"))
-
-
-# --- release check ---
-
-
-@release.command("check")
-@click.option(
-    "--since", "since_tag", default=None, help="Compare against this tag instead of auto-detecting."
-)
-@click.option(
-    "--force", is_flag=True, default=False, help="Downgrade most fatal checks to warnings."
-)
-def release_check(since_tag: str | None, force: bool):
-    """Run all pre-release checks.
-
-    \b
-    Runs 11 verification gates: version-sync, git-clean, git-tag,
-    schema-diff, schema-bump, schema-version, persistence-check, submodule,
-    generate, lint, and changelog.  Fatal failures block the
-    release unless --force is used.
-    """
-    from bootstrap.release.check import CheckSeverity
-    from bootstrap.release.check import run_all_checks
-
-    report = run_all_checks(force=force, since_tag=since_tag)
-
-    # Print results
-    click.echo("")
-    click.echo(styled([Style.BRIGHT], "=" * 60))
-    click.echo(styled([Style.BRIGHT], "Pre-Release Check Report"))
-    click.echo(styled([Style.BRIGHT], "=" * 60))
-    click.echo("")
-
-    width = max(len(r.name) for r in report.results) + 2
-
-    for r in report.results:
-        name_padded = r.name.ljust(width)
-        if r.passed:
-            click.echo(
-                styled([Fore.GREEN], f"  {r.icon} ")
-                + styled([Style.BRIGHT], name_padded)
-                + styled([Fore.GREEN], r.message)
-            )
-        elif r.severity == CheckSeverity.FATAL:
-            click.echo(
-                styled([Fore.RED], f"  {r.icon} ")
-                + styled([Style.BRIGHT], name_padded)
-                + styled([Fore.RED], r.message)
-            )
-        elif r.severity == CheckSeverity.WARN:
-            click.echo(
-                styled([Fore.YELLOW], f"  {r.icon} ")
-                + styled([Style.BRIGHT], name_padded)
-                + styled([Fore.YELLOW], r.message)
-            )
-        else:
-            click.echo(
-                styled([Fore.CYAN], f"  {r.icon} ")
-                + styled([Style.BRIGHT], name_padded)
-                + styled([Fore.CYAN], r.message)
-            )
-
-        if r.details:
-            for line in r.details.split("\n"):
-                click.echo(" " * (14 + width) + line)
-
-    click.echo("")
-    fatal = report.fatal_failures
-    warns = report.warnings
-    passed = len(report.results) - len(fatal) - len(warns)
-
-    click.echo(
-        styled([Style.BRIGHT], f"  {passed} passed, ")
-        + styled([Fore.YELLOW, Style.BRIGHT], f"{len(warns)} warnings, ")
-        + styled([Fore.RED, Style.BRIGHT], f"{len(fatal)} failures")
-    )
-    click.echo("")
-
-    if report.has_fatal_failure:
-        click.echo(styled([Fore.RED, Style.BRIGHT], "Release blocked by fatal check failures."))
-        if not force:
-            click.echo(
-                styled([Fore.YELLOW], "Re-run with --force to downgrade to warnings, ")
-                + "or fix the issues above."
-            )
-        exit(1)
-
-    if warns:
-        click.echo(
-            styled(
-                [Fore.YELLOW, Style.BRIGHT], f"Release would proceed with {len(warns)} warning(s)."
-            )
-        )
-    else:
-        click.echo(styled([Fore.GREEN, Style.BRIGHT], "All checks passed — ready to release!"))
-
-
-# --- release commit ---
-
-
-@release.command("commit")
-@click.option(
-    "--no-edit",
-    is_flag=True,
-    default=False,
-    help="Use the default message without opening an editor.",
-)
-@click.option(
-    "--dry-run", is_flag=True, default=False, help="Print the commands without executing."
-)
-def release_commit(no_edit: bool, dry_run: bool):
-    """Commit staged changes and create a git tag locally.
-
-    \b
-    Reads the version from efa.config.toml and:
-      1. Commits staged changes with message "chore: release v{version}"
-      2. Creates an annotated tag "v{version}"
-
-    By default, both the commit and tag open $EDITOR for message review.
-    Use --no-edit to accept the default messages without review.
-    Use --dry-run to preview without executing.
-
-    Does NOT push — you must push manually.
-    """
-    from bootstrap.release.git_util import check_tag_exists
-    from bootstrap.release.version import load_version
-
-    v = load_version()
-    tag = v.render_tag()
-    commit_msg = f"chore: release {tag}"
-
-    if check_tag_exists(tag):
-        raise click.ClickException(
-            f"Tag {tag} already exists. Delete it first with `git tag -d {tag}` if you want to re-tag."
-        )
-
-    # Build git commands
-    commit_cmd = ["git", "commit", "-m", commit_msg]
-    if not no_edit:
-        commit_cmd.append("--edit")
-    commit_cmd.append("-s")
-
-    tag_cmd = ["git", "tag", "-a", tag, "-m", f"Release {tag}"]
-    if not no_edit:
-        tag_cmd.append("--edit")
-
-    if dry_run:
-        click.echo(styled([Style.BRIGHT, Fore.CYAN], "[DRY-RUN] Would execute:"))
-        click.echo(f"  {' '.join(commit_cmd)}")
-        click.echo(f"  {' '.join(tag_cmd)}")
-        return
-
-    # Commit
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Committing: ") + commit_msg)
-    proc = subprocess.run(commit_cmd, cwd=PROJECT_ROOT)
-    if proc.returncode != 0:
-        raise click.ClickException(f"git commit failed with exit code {proc.returncode}")
-
-    # Tag
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Tagging: ") + tag)
-    proc = subprocess.run(tag_cmd, cwd=PROJECT_ROOT)
-    if proc.returncode != 0:
-        raise click.ClickException(f"git tag failed with exit code {proc.returncode}")
-
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Committed and tagged locally. ")
-        + styled([Fore.YELLOW], "Push manually when ready."),
-    )
-    click.echo(f"  git push origin dev && git push origin {tag}")
-
-
-# --- release changelog ---
-
-
-@release.group("changelog", cls=ClickAliasedGroup)
-def release_changelog():
-    """Changelog generation — full file and per-version documents."""
-
-
-@release_changelog.command("generate")
-def release_changelog_generate():
-    """Regenerate CHANGELOG.md using git-cliff.
-
-    Prepends a new version entry for the current version from efa.config.toml.
-    """
-    from bootstrap.release.changelog_gen import generate_full
-    from bootstrap.release.version import load_version
-
-    v = load_version()
-    tag = v.render_tag()
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Generating changelog for ")
-        + styled([Style.BRIGHT], tag)
-    )
-    generate_full(v)
-    click.echo(styled([Fore.GREEN], "  Prepended entry to CHANGELOG.md"))
-
-
-@release_changelog.command("detail")
-@click.option(
-    "--no-edit",
-    is_flag=True,
-    default=False,
-    help="Write template as-is without opening editor.",
-)
-def release_changelog_detail(no_edit: bool):
-    """Generate bi-lingual version announcements for in-app release notes.
-
-    By default, opens $EDITOR with a template containing en-us/zh-cn summary sections.
-    Use --no-edit to write the generated template as-is without manual editing.
-    On save (or if --no-edit), writes authored .md files to assets/content/announcements/{en,zh}/.
-    """
-    from bootstrap.release.changelog_gen import generate_detail
-    from bootstrap.release.version import load_version
-
-    v = load_version()
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Preparing version announcements for ")
-        + styled([Style.BRIGHT], v.render_semver())
-    )
-    generate_detail(v, no_edit=no_edit)
-    click.echo(styled([Fore.GREEN], "  Written to assets/content/announcements/"))
-
-
-@cli.group(cls=ClickAliasedGroup)
-def etc():
-    """Extra toolsets."""
-
-
-@etc.command("codeart")
-def etc_codeart_cmd():
-    """Generate the codeart image."""
-    warning("Generate codeart requires tokei with json output installed and exported via path.")
-    tokei = get_command("tokei")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "tokei . -o json")
-    stdout = __execute_command([tokei, ".", "-o", "json"], "TOKEI OUTPUT")
-
-    output_file = PROJECT_ROOT / "codeart.png"
-    generate_codeart(stdout, output_file)
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Codeart image generated successfully: ")
-        + str(output_file)
-    )
-
-
-@etc.group(cls=ClickAliasedGroup)
-def site():
-    """Landing page site commands."""
-
-
-@site.command("dev")
-def site_dev():
-    """Start the SvelteKit dev server."""
-    pnpm = get_command("pnpm")
-    __execute_command([pnpm, "--filter", "efa-tech", "dev"], "SITE DEV", live_stdout=True)
-
-
-@site.command("build")
-def site_build():
-    """Build the static site for Cloudflare Pages."""
-    pnpm = get_command("pnpm")
-    __execute_command([pnpm, "--filter", "efa-tech", "build"], "SITE BUILD", live_stdout=True)
-
-
-@cli.group(aliases=["t"], cls=ClickAliasedGroup)
-def test():
-    """Run project test suites."""
-
-
-@test.command("python")
-def test_python():
-    """Run Python tests via pytest."""
-    uv = get_command("uv")
-    click.echo(
-        styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "uv run pytest bootstrap/tests/"
-    )
-    __execute_command([uv, "run", "pytest", "bootstrap/tests/"], "PYTEST OUTPUT")
-
-
-@test.command("dart")
-def test_dart():
-    """Run Flutter/Dart tests."""
-    flutter = get_command("flutter")
-    click.echo(styled([Style.BRIGHT, Fore.GREEN], "Executing command: ") + "flutter test")
-    __execute_command([flutter, "test"], "FLUTTER TEST OUTPUT")
-
-
-@test.command("all")
-def test_all():
-    """Run all test suites (Python + Dart)."""
-    ctx = click.get_current_context()
-    ctx.invoke(test_python)
-    click.echo()
-    ctx.invoke(test_dart)
-
-
-register_ci_commands(cli)
+register_all_commands(cli)
 
 
 cli()
